### 2020-02-25 #2114 @jordanperr
Hi @kyleniemeyer and @sjpfenninger,

It looks like most of the Whedon checks were passed. Would it be preferable to swap out the URL for the DOI in reference "Understanding Biases in Pre-Construction Estimate"? It would be no trouble for me to do. I'll monitor this thread for any further questions or changes.

Thank you, and I look forward to your review.
