### 2020-02-27 #2114 @kyleniemeyer
@jordanperr we can generate the PDF from a branch using this command:

`@whedon generate pdf from branch custom-branch-name`
