+++ paper/paper.bib
@@ -40,7 +40,7 @@ year = {2018}
   volume={1037},
   number={6},
   pages={062009},
-  url={http://stacks.iop.org/1742-6596/1037/i=6/a=062009},
+  url={https://doi.org/10.1088/1742-6596/1037/6/062009},
   year={2018},
   abstract={The pre-construction energy generation of a wind farm (P50) is difficult to estimate and evaluate. This paper presents a methodology to measure the accuracy of the p50 prediction, which we call the Historical Validation Survey (HVS), for several wind farms in the continental United States. Our results indicate that there is a bias between predicted and measured energy, even when controlling for factors like grid curtailment and resource variability. We also find that our results depend on the assumptions we make during analysis, which we quantify with a sensitivity analysis. This method allows the estimation of uncertainty we have in our findings. When we account for reasonable ranges of model assumptions, we find that, in the most optimistic case, there is still a bulk −5.5% bias when estimating pre-construction energy generation. When controlling for grid curtailment this number reduces to a range of −3.5 to −4.5%.}
 }
