### 2020-03-23 #2114 @jordanperr
@sjpfenninger Hi Stefan, I hope all is well with you during these trying times. I'd like to know what kind of progress we should expect on this article over the next couple of weeks. I will be here to answer any questions or provide revisions. Thank you, and stay safe.
