### 2020-03-31 #2171 @whedon
[REVIEW]: OpenOA: An Open-Source Codebase For Operational Analysis of Wind Farms
**Submitting author:** @jordanperr (<a href="http://orcid.org/0000-0003-1571-1887">Jordan Perr-Sauer</a>)
**Repository:** <a href="https://github.com/NREL/OpenOA" target ="_blank">https://github.com/NREL/OpenOA</a>
**Version:** v2.1
