### 2020-03-31 #2171 @gschivley
@sjpfenninger how strict is JOSS about requiring OSI approved licenses? It looks like the authors use a  modified version of the [BSD 3-clause](https://opensource.org/licenses/BSD-3-Clause). There are very minor changes to wording in the original 3 clauses, plus an extra clause:
>3. The entire corresponding source code of any redistribution, with or without modification, by a research entity,
including but not limited to any contracting manager/operator of a United States National Laboratory, any
institution of higher learning, and any non-profit organization, must be made publicly available under this license
for as long as the redistribution is made available by the research entity.
