### 2020-03-31 #2171 @danielskatz
We are very strict about this!  We will not publish this submission with a modified license
