### 2020-03-31 #2171 @gschivley
Thanks @danielskatz. Does that include minor modifications to the original clauses like those below?
> 1. Redistributions of source code must retain the above copyright notice, **the above government rights notice**, this list of conditions and the following disclaimer.

> The name of the copyright holder, contributors, **the United States Government, the United States Department of Energy, or any of their employees** may not be used to endorse or promote products derived from this software without specific prior written permission.

NREL has other repos that use a standard BSD 3-Clause so I would think they should be able to change this one, I just want to be clear on exactly how strict JOSS is about deviations in wording.
