### 2020-03-31 #2171 @danielskatz
No deviations in wording are allowed. Any deviation is a different (and thus non-OSI-approved) license.
