### 2020-03-31 #2171 @jordanperr
@danielskatz @gschivley - I will bring this up with our team and see if we can accommodate this request. I imagine this license issue may have come up in the past for JOSS if they've published any work that was funded by the US government.
