### 2020-03-31 #2171 @danielskatz
In other cases that I remember, labs have found a way to use a standard license.
