+++ operational_analysis/methods/electrical_losses_uq.py
@@ -1,254 +1,0 @@
-# This class defines key analytical routines for calculating electrical losses for 
-# a wind plant using operational data. Electrical loss is calculated per month and on 
-# an average annual basis by comparing monthly energy production from the turbines
-# and the revenue meter
-
-from operational_analysis import logged_method_call
-from operational_analysis import logging
-import numpy as np
-from tqdm import tqdm
-
-logger = logging.getLogger(__name__)
-
-class ElectricalLosses(object):
-    """
-    A serial (Pandas-driven) implementation of calculating the average monthly and annual
-    electrical losses at a wind plant, and their uncertainty. 
-    Energy output from the turbine SCADA meter and the wind plant revenue 
-    meter are used to estimate electrical losses.
-    
-    The approach is to first calculate daily sums of turbine and revenue meter energy over the 
-    plant period of record. Only those days where all turbines and the revenue meter were 
-    reporting for all timesteps are considered. Electrical loss is then the difference in 
-    total turbine energy production and meter production over those concurrent days. 
-    
-    A Monte Carlo approach is applied to sample revenue meter data and SCADA data
-    with a 0.5% imposed uncertainty, and one filtering parameter is sampled too.
-    The uncertainty in estimated electrical losses is quantified as standard deviation
-    of the distribution of losses obtained from the MC sampling.
-    
-    In the case that meter data is not provided on a daily or sub-daily basis (e.g. monthly), a
-    different approach is implemented. The sum of daily turbine energy is corrected for any missing 
-    reported energy data from the turbines based on the ratio of expected number of data counts per day 
-    to the actual. Daily corrected sum of turbine energy is then summed on a monthly basis. Electrical 
-    loss is then the difference between total corrected turbine energy production and meter production 
-    over those concurrent months.   
-    """ 
-
-    @logged_method_call
-    def __init__(self, plant, num_sim, uncertainty_meter=0.005, uncertainty_scada=0.005,
-                 uncertainty_correction_thresh=(0.9,0.97)):
-        """
-        Initialize electrical losses class with input parameters
-
-        Args:
-         plant(:obj:`PlantData object`): PlantData object from which EYAGapAnalysis should draw data.
-
-        """
-        logger.info("Initializing Electrical Losses Object")
-        
-        self._plant = plant
-        self.num_sim = num_sim
-        
-        self._time_conversion = {'1T': 1.,
-                                 '5T': 5.,
-                                 '10T': 10.,
-                                 '30T': 30.,
-                                 '1H': 60.,
-                                 'D': 60 * 24}
-        
-        self._min_per_hour = 60 # Mintues per hour converter
-        self._hours_per_day= 24 # Hours per day converter
-    
-        # Define relevant uncertainties, to be applied in Monte Carlo sampling
-        self.uncertainty_meter = np.float64(uncertainty_meter)
-        self.uncertainty_scada = np.float64(uncertainty_scada)
-        self.uncertainty_correction_thresh = np.array(uncertainty_correction_thresh, dtype=np.float64)  
-
-
-    @logged_method_call
-    def run(self):
-        """
-        Run the electrical loss calculation in order by calling this function.
-        
-        Args:
-            (None)
-            
-        Returns:
-            (None)
-        """
-        
-        # Process SCADA data to daily sums
-        self.process_scada()
-        
-        # Process meter data to daily sums (if time frequency is less than monthly)
-        self._monthly_meter = True # Keep track of reported meter data frequency
-        
-        if (self._plant._meter_freq != 'MS') & (self._plant._meter_freq != 'M') & (self._plant._meter_freq != '1MS'):
-            self.process_meter()
[…]

[…]
