### 2020-04-03 #2171 @jordanperr
@gschivley - Great news! I was able to change the license to a standard BSD license. This change has been merged in to the master branch of the OpenOA repository.
