### 2020-04-06 #2171 @jordanperr
Yes. I tried to make this author list a most complete reflection of the NREL team. Some of those authors may only have contributed in the past, using an internal git server, or they have been a guiding force in planning or architecture of the software.
