### 2020-04-08 ec56304
Delete plant_analysis_github.py
