### 2020-04-13 #2171 @jordanperr
Updated spelling of an author's name and added another author to the paper.
