### 2020-04-17 8a7c88d
Added tests and addressed Eric's comments
Two tests have been added. Also, both the TIE method and the electrical loss method now have an option to let the user decide whether to include UQ or not.
