+++ operational_analysis/methods/turbine_long_term_gross_energy_uq.py
@@ -47,7 +47,7 @@ class TurbineLongTermGrossEnergy(object):
     """
 
     @logged_method_call
-    def __init__(self, plant, num_sim = 100, uncertainty_meter=0.005, uncertainty_scada=0.005,
+    def __init__(self, plant, UQ = 'Y', num_sim = 100, uncertainty_scada=0.005,
                   uncertainty_wind_bin_thresh=(1, 3), uncertainty_max_power_filter=(0.8, 0.9), 
                   uncertainty_correction_threshold=(0.85, 0.95)):
     
@@ -55,8 +55,8 @@ class TurbineLongTermGrossEnergy(object):
         Initialize turbine long-term gross energy analysis with data and parameters.
         Args:
          plant(:obj:`PlantData object`): PlantData object from which TurbineLongTermGrossEnergy should draw data.
+         UQ:(:obj:`str`): choice whether to perform ('Y') or not ('N') uncertainty quantification
          num_sim:(:obj:`int`): number of Monte Carlo simulations
-         uncertainty_meter(:obj:`float`): uncertainty imposed to revenue meter data
          uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
          uncertainty_max_power_filter(:obj:`float`): Maximum power threshold (fraction) to which the bin filter 
                                             should be applied (default 0.85)
@@ -65,6 +65,17 @@ class TurbineLongTermGrossEnergy(object):
                                                 should be corrected (default is 0.90)
         """
         logger.info("Initializing TurbineLongTermGrossEnergy Object")
+        
+        # Check that selected UQ is allowed
+        if UQ == 'Y':
+            logger.info("Note: uncertainty quantification will be performed in the calculation")
+            self.num_sim = num_sim
+        elif UQ == 'N':    
+            logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
+            self.num_sim = 1
+        else:   
+            raise ValueError("UQ has to either be Y (uncertainty quantification performed, default) or N (uncertainty quantification NOT performed)")
+        self.UQ = UQ
         
         self._plant = plant  # Set plant as attribute of analysis object
         self._turbs = self._plant._scada.df['id'].unique() # Store turbine names
@@ -81,7 +92,7 @@ class TurbineLongTermGrossEnergy(object):
         self._model_results = {}
         self._turb_lt_gross = {}
         self._scada_daily_valid = pd.DataFrame()
-              
+
         # Dictionary to convert time interval indicator into minutes
         self._time_conversion = {'10T': 10.,
                                  '5T': 5.,
@@ -89,9 +100,6 @@ class TurbineLongTermGrossEnergy(object):
         
         # Set number of 'valid' counts required when summing data to daily values
         self._num_valid_daily = 60. / self._time_conversion[self._plant._scada_freq] * 24
-        
-        # Number of times Monte Carlo sampling is performed
-        self.num_sim = num_sim
         
         # Define relevant uncertainties, to be applied in Monte Carlo sampling
         self.uncertainty_scada = np.float64(uncertainty_scada)
@@ -125,7 +133,8 @@ class TurbineLongTermGrossEnergy(object):
         self.plot_dir = plot_dir
         
         # Setup and run Monte Carlo
-        self.setup_TIE_monte_carlo_inputs()
+        if self.UQ == 'Y':
+            self.setup_TIE_monte_carlo_inputs()
         self.run_TIE_monte_carlo()
         
         # Log the completion of the run
@@ -169,7 +178,10 @@ class TurbineLongTermGrossEnergy(object):
             turb_capac = dic[t].wtur_W_avg.max()
             
             # Monte Carlo sample max_power_filter
[…]
