+++ operational_analysis/methods/plant_analysis_v2.py
@@ -1,0 +1,1088 @@
+# PrufPlantAnalysis
+#
+# This class defines key analytical routines for the PRUF/WRA Benchmarking
+# standard operational assessment.
+#
+# The PrufPlantAnalysis object is a factory which instantiates either the Pandas, Dask, or Spark
+# implementation depending on what the user prefers.
+#
+# The resulting object is loaded as a plugin into each PlantData object.
+
+import random
+import numpy as np
+import pandas as pd
+import matplotlib
+matplotlib.use('Agg')
+import statsmodels.api as sm
+from tqdm import tqdm
+from sklearn.linear_model import LinearRegression
+from sklearn.metrics import r2_score
+from sklearn.metrics import mean_squared_error
+from operational_analysis.toolkits import met_data_processing as mt
+from operational_analysis.toolkits import timeseries as tm
+#from operational_analysis.toolkits.machine_learning_setup import MachineLearningSetup
+from operational_analysis.toolkits import unit_conversion as un
+from operational_analysis.types import timeseries_table
+from operational_analysis import logged_method_call
+from operational_analysis import logging
+
+logger = logging.getLogger(__name__)
+
+
+class MonteCarloAEP(object):
+    """
+    A serial (Pandas-driven) implementation of the benchmark PRUF operational
+    analysis implementation. This module collects standard processing and
+    analysis methods for estimating plant level operational AEP and uncertainty.
+
+    The preprocessing should run in this order:
+
+        1. Process revenue meter energy - creates monthly/daily data frame, gets revenue meter on monthly/daily basis, and adds
+           data flag
+        2. Process loss estimates - add monthly/daily curtailment and availabilty losses to monthly/daily data frame
+        3. Process reanalysis data - add monthly/daily density-corrected wind speeds, temperature (if used) and wind direction (if used)
+            from several reanalysis products to the monthly data frame
+        4. Set up Monte Carlo - create the necessary Monte Carlo inputs to the OA process
+        5. Run AEP Monte Carlo - run the OA process iteratively to get distribution of AEP results
+
+    The end result is a distribution of AEP results which we use to assess expected AEP and associated uncertainty
+    """
+
+    @logged_method_call
+    def __init__(self, plant, uncertainty_meter=0.005, uncertainty_losses=0.05,
+                 uncertainty_loss_max=(10, 20), uncertainty_windiness=(10, 20), uncertainty_outlier=(2, 3.1),
+                 uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', 
+                 reg_temperature = 'N', reg_winddirection = 'N'):
+        """
+        Initialize APE_MC analysis with data and parameters.
+
+        Args:
+         plant(:obj:`PlantData object`): PlantData object from which PlantAnalysis should draw data.
+         uncertainty_meter(:obj:`float`): uncertainty on revenue meter data
+         uncertainty_losses(:obj:`float`): uncertainty on long-term losses
+         uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
+         uncertainty_windiness(:obj:`float`): number of years to use for the windiness correction
+         uncertainty_outlier(:obj:`float`): the Huber parameter which controls sensitivity of outlier detection in robust linear regression
+         uncertainty_nan_energy(:obj:`float`): threshold to flag days/months based on NaNs
+         time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
+         reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr')
[…]
