+++ operational_analysis/methods/plant_analysis_v3.py
@@ -18,10 +18,12 @@ from tqdm import tqdm
 from sklearn.linear_model import LinearRegression
 from sklearn.metrics import r2_score
 from sklearn.metrics import mean_squared_error
+from sklearn.model_selection import KFold
 from operational_analysis.toolkits import met_data_processing as mt
 from operational_analysis.toolkits import timeseries as tm
-#from operational_analysis.toolkits.machine_learning_setup import MachineLearningSetup
+from operational_analysis.toolkits.machine_learning_setup import MachineLearningSetup
 from operational_analysis.toolkits import unit_conversion as un
+from operational_analysis.toolkits import filters
 from operational_analysis.types import timeseries_table
 from operational_analysis import logged_method_call
 from operational_analysis import logging
@@ -50,7 +52,8 @@ class MonteCarloAEP(object):
 
     @logged_method_call
     def __init__(self, plant, uncertainty_meter=0.005, uncertainty_losses=0.05,
-                 uncertainty_loss_max=(10, 20), uncertainty_windiness=(10, 20), uncertainty_outlier=(2, 3.1),
+                 uncertainty_windiness=(10, 20), uncertainty_wind_bin_thresh=(1, 3), 
+                 uncertainty_loss_max=(10, 20), uncertainty_max_power_filter=(0.8, 0.9), 
                  uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', 
                  reg_temperature = 'N', reg_winddirection = 'N'):
         """
@@ -60,9 +63,11 @@ class MonteCarloAEP(object):
          plant(:obj:`PlantData object`): PlantData object from which PlantAnalysis should draw data.
          uncertainty_meter(:obj:`float`): uncertainty on revenue meter data
          uncertainty_losses(:obj:`float`): uncertainty on long-term losses
-         uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
          uncertainty_windiness(:obj:`float`): number of years to use for the windiness correction
-         uncertainty_outlier(:obj:`float`): the Huber parameter which controls sensitivity of outlier detection in robust linear regression
+         uncertainty_wind_bin_thresh(:obj:`float`): The filter threshold for each bin (default is 2 m/s)
+         uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
+         uncertainty_max_power_filter(:obj:`float`): Maximum power threshold (fraction) to which the bin filter 
+                                            should be applied (default 0.85)
          uncertainty_nan_energy(:obj:`float`): threshold to flag days/months based on NaNs
          time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
          reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr')
@@ -76,19 +81,15 @@ class MonteCarloAEP(object):
         self._plant = plant  # defined at runtime
 
         # Memo dictionaries help speed up computation
-        self.long_term_sampling = {}  # Combinations of long-term reanalysis data sampling
-        self.long_term_sampling_wd = {} 
-        self.long_term_sampling_wd_sin = {} 
-        self.long_term_sampling_wd_cos = {} 
-        self.long_term_sampling_temp = {} 
         self.outlier_filtering = {}  # Combinations of outlier filter results
 
         # Define relevant uncertainties, data ranges and max thresholds to be applied in Monte Carlo sampling
         self.uncertainty_meter = np.float64(uncertainty_meter)
         self.uncertainty_losses = np.float64(uncertainty_losses)
[…]
