### 2020-05-14 3c20f76
Add augmented AEP codes
