### 2020-05-21 bc749c2
adding github-actions ci workflow
