### 2020-05-21 9dc7e08
fixing syntax error in python-ci
