+++ .github/workflows/python-ci.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements-python3.txt; fi
+        pip install -r requirements-python3.txt
         yes | unzip examples/operational_AEP_analysis/data/eia_example_data.zip -d examples/operational_AEP_analysis/data/
         yes | unzip examples/turbine_analysis/data/example_20180829.zip -d examples/turbine_analysis/data/
         cp examples/turbine_analysis/data/example_20180829/scada_10min_4cols.csv examples/turbine_analysis/data/scada_10min_4cols.csv
