### 2020-05-21 b3d95a1
upgrading statsmodels in requirements file to try and avoid issue with numpy during installatin
