+++ requirements.txt
@@ -1,14 +1,16 @@
 # Requirements for the OpenOA software package
-numpy==1.14.0
-scipy==1.0.0
-pandas==0.20.3
-geopandas==0.3.0
-pygam==0.4.0
-tqdm==4.19.5
-statsmodels==0.8.0
+eia-python==1.22
+geopandas==0.4.0
+numpy==1.15.4
+pandas==0.23.4
+pygam==0.8.0
+scipy==1.1.0
+statsmodels==0.11
+tqdm==4.28.1
+
+# Not included in docs requirements
 pytest==3.4.0
 pytest-cov==2.5.1
-scikit_learn==0.18.1
+scikit_learn==0.20.1
 matplotlib==2.1.0
-eia-python==1.22
 requests==2.21.0
