### 2020-05-21 d449585
updated pytest requirement
