+++ requirements.txt
@@ -9,8 +9,8 @@ statsmodels==0.11
 tqdm==4.28.1
 
 # Not included in docs requirements
-pytest==3.4.0
-pytest-cov==2.5.1
+pytest==5.4.2
+pytest-cov==2.8.1
 scikit_learn==0.20.1
 matplotlib==2.1.0
 requests==2.21.0
