### 2020-05-21 72fa449
updated packages in requirements to latest versions, fixed deprecated as_matrix call, deleted travis yml
