+++ requirements.txt
@@ -1,16 +1,16 @@
 # Requirements for the OpenOA software package
-eia-python==1.22
-geopandas==0.4.0
-numpy==1.15.4
-pandas==0.23.4
-pygam==0.8.0
-scipy==1.1.0
-statsmodels==0.11
-tqdm==4.28.1
+eia-python>=1.22
+geopandas>=0.4.0
+numpy>=1.15.4
+pandas>=0.23.4
+pygam>=0.8.0
+scipy>=1.1.0
+statsmodels>=0.11
+tqdm>=4.28.1
 
 # Not included in docs requirements
-pytest==5.4.2
-pytest-cov==2.8.1
-scikit_learn==0.20.1
-matplotlib==2.1.0
-requests==2.21.0
+pytest>=5.4.2
+pytest-cov>=2.8.1
+scikit_learn>=0.20.1
+matplotlib>=2.1.0
+requests>=2.21.0
