### 2020-05-21 0ce39df
changing name of workflow. Adding schedule
