+++ .github/workflows/python-ci.yml
@@ -1,10 +1,12 @@
-name: Test the python package
+name: Tests
 
 on:
   push:
     branches: [ master, develop ]
   pull_request:
     branches: [ master, develop ]
+  schedule:
+    branches: [ develop ]
 
 jobs:
   build:
@@ -24,9 +26,11 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
+    - name: Unzip example data
+      run: |
         yes | unzip examples/operational_AEP_analysis/data/eia_example_data.zip -d examples/operational_AEP_analysis/data/
         yes | unzip examples/turbine_analysis/data/example_20180829.zip -d examples/turbine_analysis/data/
         cp examples/turbine_analysis/data/example_20180829/scada_10min_4cols.csv examples/turbine_analysis/data/scada_10min_4cols.csv
     - name: Test with pytest
       run: |
-        python setup.py test
+        pytest -o python_files=test/*.py --cov=operational_analysis
