### 2020-05-21 c6271a7
changing name of workflow. Adding schedule
