+++ .github/workflows/python-ci.yml
@@ -5,8 +5,6 @@ on:
     branches: [ master, develop ]
   pull_request:
     branches: [ master, develop ]
-  schedule:
-    branches: [ develop ]
 
 jobs:
   build:
