### 2020-05-21 6a62557
decided not to tackle geopandas today. updated readme
