+++ setup.py
@@ -72,8 +72,8 @@ setup(name='OpenOA',
       install_requires=["numpy",
                         "scipy",
                         "pandas",
-                        "geopandas",
                         "pygam",
+                        "geopandas",
                         "tqdm",
                         "statsmodels",
                         "scikit_learn",
