### 2020-05-21 9ce78b0
Updating PR tests to make them run faster
