### 2020-05-22 c54936d
Incorporating some of Jordan's initial comments
