### 2020-05-22 47be19d
Removing initial version of scripts, unifying with the new ones
