+++ operational_analysis/methods/electrical_losses.py
@@ -5,19 +5,27 @@
 
 from operational_analysis import logged_method_call
 from operational_analysis import logging
+import numpy as np
+from tqdm import tqdm
 
 logger = logging.getLogger(__name__)
 
 class ElectricalLosses(object):
     """
     A serial (Pandas-driven) implementation of calculating the average monthly and annual
-    electrical losses at a wind plant. Energy output from the turbine SCADA meter and the 
-    wind plant revenue meter are used to estimate electrical losses.
+    electrical losses at a wind plant, and their uncertainty. 
+    Energy output from the turbine SCADA meter and the wind plant revenue 
+    meter are used to estimate electrical losses.
     
     The approach is to first calculate daily sums of turbine and revenue meter energy over the 
     plant period of record. Only those days where all turbines and the revenue meter were 
     reporting for all timesteps are considered. Electrical loss is then the difference in 
     total turbine energy production and meter production over those concurrent days. 
+    
+    A Monte Carlo approach is applied to sample revenue meter data and SCADA data
+    with a 0.5% imposed uncertainty, and one filtering parameter is sampled too.
+    The uncertainty in estimated electrical losses is quantified as standard deviation
+    of the distribution of losses obtained from the MC sampling.
     
     In the case that meter data is not provided on a daily or sub-daily basis (e.g. monthly), a
     different approach is implemented. The sum of daily turbine energy is corrected for any missing 
@@ -28,15 +36,34 @@ class ElectricalLosses(object):
     """ 
 
     @logged_method_call
-    def __init__(self, plant, correction_thresh=.95):
+    def __init__(self, plant, UQ = 'Y', num_sim = 20000, uncertainty_meter=0.005, uncertainty_scada=0.005,
+                 uncertainty_correction_thresh=(0.9,0.995), correction_thresh=.95):
         """
         Initialize electrical losses class with input parameters
 
         Args:
          plant(:obj:`PlantData object`): PlantData object from which EYAGapAnalysis should draw data.
-
+         num_sim:(:obj:`int`): number of Monte Carlo simulations
+         UQ:(:obj:`str`): choice whether to perform ('Y') or not ('N') uncertainty quantification
+         uncertainty_meter(:obj:`float`): uncertainty imposed to revenue meter data
+         uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
+         uncertainty_correction_threshold(:obj:`float`): The interval of data availability thresholds (fractions) 
+                                                         under which months should be eliminated - used for UQ cases    
+         correction_thresh(:obj:`float`): The data availability threshold (fraction) under which months 
+                                         should be eliminated (default is 0.95)  - used for not-UQ cases                                          
         """
         logger.info("Initializing Electrical Losses Object")
+        
+        # Check that selected UQ is allowed
+        if UQ == 'Y':
+            logger.info("Note: uncertainty quantification will be performed in the calculation")
+            self.num_sim = num_sim
+        elif UQ == 'N':    
+            logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
+            self.num_sim = 1
+        else:   
+            raise ValueError("UQ has to either be Y (uncertainty quantification performed, default) or N (uncertainty quantification NOT performed)")
+        self.UQ = UQ
         
         self._plant = plant
         
@@ -49,6 +76,11 @@ class ElectricalLosses(object):
         
         self._min_per_hour = 60 # Mintues per hour converter
         self._hours_per_day= 24 # Hours per day converter
+    
+        # Define relevant uncertainties, to be applied in Monte Carlo sampling
+        self.uncertainty_meter = np.float64(uncertainty_meter)
[…]
