### 2020-05-22 5f0b87c
Update versions of dependencies. Switch to Github Actions. (#67)
* adding github-actions ci workflow

* fixing syntax error in python-ci

* upgrading statsmodels in requirements file to try and avoid issue with numpy during installatin

* updated pytest requirement

* updated packages in requirements to latest versions, fixed deprecated as_matrix call, deleted travis yml

* changing name of workflow. Adding schedule

* changing name of workflow. Adding schedule

* decided not to tackle geopandas today. updated readme

* adding nicolas name to readme
