+++ setup.py
@@ -72,15 +72,15 @@ setup(name='OpenOA',
       install_requires=["numpy",
                         "scipy",
                         "pandas",
-                        "geopandas",
                         "pygam",
+                        "geopandas",
                         "tqdm",
                         "statsmodels",
                         "scikit_learn",
                         "EIA-python",
                         "requests"],
       tests_require=['pytest', 'pytest-cov'],
-      python_requires='>=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*,!=3.5.*',
+      python_requires='>=3.6',
       cmdclass={'test': PyTest, 'integrate':PyTestIntegrate, 'unit':PyTestUnit},
       license='None'
       )
