+++ operational_analysis/methods/turbine_long_term_gross_energy.py
@@ -47,7 +47,7 @@ class TurbineLongTermGrossEnergy(object):
     """
 
     @logged_method_call
-    def __init__(self, plant, UQ = 'Y', num_sim = 100, uncertainty_scada=0.005,
+    def __init__(self, plant, UQ = True, num_sim = 100, uncertainty_scada=0.005,
                   uncertainty_wind_bin_thresh=(1, 3), uncertainty_max_power_filter=(0.8, 0.9), 
                   uncertainty_correction_threshold=(0.85, 0.95)):
     
@@ -55,26 +55,26 @@ class TurbineLongTermGrossEnergy(object):
         Initialize turbine long-term gross energy analysis with data and parameters.
         Args:
          plant(:obj:`PlantData object`): PlantData object from which TurbineLongTermGrossEnergy should draw data.
-         UQ:(:obj:`str`): choice whether to perform ('Y') or not ('N') uncertainty quantification
+         UQ:(:obj:`bool`): choice whether to perform ('Y') or not ('N') uncertainty quantification
          num_sim:(:obj:`int`): number of Monte Carlo simulations
          uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
-         uncertainty_max_power_filter(:obj:`float`): Maximum power threshold (fraction) to which the bin filter 
+         uncertainty_max_power_filter(:obj:`tuple`): Maximum power threshold (fraction) to which the bin filter 
                                             should be applied (default 0.85)
-         uncertainty_wind_bin_thresh(:obj:`float`): The filter threshold for each bin (default is 2 m/s)
-         uncertainty_correction_threshold(:obj:`float`): The threshold (fraction) above which daily scada energy data
+         uncertainty_wind_bin_thresh(:obj:`tuple`): The filter threshold for each bin (default is 2 m/s)
+         uncertainty_correction_threshold(:obj:`tuple`): The threshold (fraction) above which daily scada energy data
                                                 should be corrected (default is 0.90)
         """
         logger.info("Initializing TurbineLongTermGrossEnergy Object")
         
         # Check that selected UQ is allowed
-        if UQ == 'Y':
+        if UQ == True:
             logger.info("Note: uncertainty quantification will be performed in the calculation")
             self.num_sim = num_sim
-        elif UQ == 'N':    
+        elif UQ == False:    
             logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
             self.num_sim = 1
         else:   
-            raise ValueError("UQ has to either be Y (uncertainty quantification performed, default) or N (uncertainty quantification NOT performed)")
+            raise ValueError("UQ has to either be True (uncertainty quantification performed, default) or False (uncertainty quantification NOT performed)")
         self.UQ = UQ
         
         self._plant = plant  # Set plant as attribute of analysis object
@@ -133,7 +133,7 @@ class TurbineLongTermGrossEnergy(object):
         self.plot_dir = plot_dir
         
         # Setup and run Monte Carlo
-        if self.UQ == 'Y':
+        if self.UQ == True:
             self.setup_TIE_monte_carlo_inputs()
         self.run_TIE_monte_carlo()
         
@@ -178,9 +178,9 @@ class TurbineLongTermGrossEnergy(object):
             turb_capac = dic[t].wtur_W_avg.max()
             
             # Monte Carlo sample max_power_filter
-            if self.UQ == 'Y':
+            if self.UQ == True:
                 self._max_power_filter = self._mc_max_power_filter[n]
-            elif self.UQ == 'N':
+            elif self.UQ == False:
                 self._max_power_filter = np.mean(self.uncertainty_max_power_filter)
[…]
