+++ operational_analysis/methods/electrical_losses.py
@@ -36,7 +36,7 @@ class ElectricalLosses(object):
     """ 
 
     @logged_method_call
-    def __init__(self, plant, UQ = 'Y', num_sim = 20000, uncertainty_meter=0.005, uncertainty_scada=0.005,
+    def __init__(self, plant, UQ = True, num_sim = 20000, uncertainty_meter=0.005, uncertainty_scada=0.005,
                  uncertainty_correction_thresh=(0.9,0.995), correction_thresh=.95):
         """
         Initialize electrical losses class with input parameters
@@ -44,10 +44,10 @@ class ElectricalLosses(object):
         Args:
          plant(:obj:`PlantData object`): PlantData object from which EYAGapAnalysis should draw data.
          num_sim:(:obj:`int`): number of Monte Carlo simulations
-         UQ:(:obj:`str`): choice whether to perform ('Y') or not ('N') uncertainty quantification
+         UQ:(:obj:`bool`): choice whether to perform ('Y') or not ('N') uncertainty quantification
          uncertainty_meter(:obj:`float`): uncertainty imposed to revenue meter data
          uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
-         uncertainty_correction_threshold(:obj:`float`): The interval of data availability thresholds (fractions) 
+         uncertainty_correction_threshold(:obj:`tuple`): The interval of data availability thresholds (fractions) 
                                                          under which months should be eliminated - used for UQ cases    
          correction_thresh(:obj:`float`): The data availability threshold (fraction) under which months 
                                          should be eliminated (default is 0.95)  - used for not-UQ cases                                          
@@ -55,14 +55,14 @@ class ElectricalLosses(object):
         logger.info("Initializing Electrical Losses Object")
         
         # Check that selected UQ is allowed
-        if UQ == 'Y':
+        if UQ == True:
             logger.info("Note: uncertainty quantification will be performed in the calculation")
             self.num_sim = num_sim
-        elif UQ == 'N':    
+        elif UQ == False:    
             logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
             self.num_sim = 1
         else:   
-            raise ValueError("UQ has to either be Y (uncertainty quantification performed, default) or N (uncertainty quantification NOT performed)")
+            raise ValueError("UQ has to either be True (uncertainty quantification performed, default) or False (uncertainty quantification NOT performed)")
         self.UQ = UQ
         
         self._plant = plant
@@ -106,7 +106,7 @@ class ElectricalLosses(object):
             self._monthly_meter = False # Set to false if sub-monthly frequency
         
         # Monte Carlo sampling
-        if self.UQ == 'Y':
+        if self.UQ == True:
             self.setup_monte_carlo_inputs()
         
         # Calculate electrical losses, Monte Carlo approach
@@ -234,7 +234,7 @@ class ElectricalLosses(object):
                 scada_monthly['perc'] = scada_monthly['count']/scada_monthly['expected_count_monthly']
                                 
                 # Filter out months in which there was less than x% of total running (all turbines at all timesteps)
-                if self.UQ == 'Y':
+                if self.UQ == True:
                     self._correction_thresh = self._mc_correction_thresh[n]
                     
                 scada_monthly = scada_monthly.loc[scada_monthly['perc']>= self._correction_thresh, :]
@@ -253,11 +253,11 @@ class ElectricalLosses(object):
             # Calculate electrical loss from difference of sum of turbine and meter energy 
             self._total_turbine_energy = merge_sum['turbine_energy_kwh']
             self._total_meter_energy = merge_sum['energy_kwh']
-            if self.UQ == 'Y':
+            if self.UQ == True:
[…]
