### 2020-06-03 ff8f72e
Added example data based on the ENGIE La Haute Borne project and a corresponding project class
