+++ examples/project_ENGIE.py
@@ -1,0 +1,212 @@
+#################################################
+# Data import script for La Haute Borne Project #
+#################################################
+"""
+This is the import script for the example ENGIE La Haute Borne project. Below
+is a description of data quality for each data frame and an overview of the
+steps taken to correct the raw data for use in the PRUF OA code.
+
+1. SCADA dataframe
+- 10-minute SCADA data for each of the four turbines in the project
+- Power, wind speed, wind direction, nacelle position, wind vane, temperature,
+  blade pitch
+- Corrects to UTC timezone when importing
+- Removes some outliers and stuck sensor values
+
+2. Meter data frame
+- 10-minute performance data provided in energy units (kWh)
+- Generated by adding artificial electrical loss and uncertaitny to SCADA data
+- No need for timezone correction
+
+3. Curtailment data frame
+- 10-minute availability and curtailment data in kwh
+- Generated by estimating availability and curtailment from SCADA data
+- Below-normal production classified as curtailment if present at all turbines.
+  Otherwise, classified as availability loss  
+- No need for timezone correction
+
+4. Reanalysis products
+- Import MERRA-2 and ERA5 1-hour reanalysis data
+- Wind speed, wind direction, temperature, and density
+
+"""
+from operational_analysis.types import PlantData
+
+import numpy as np
+import pandas as pd
+import operational_analysis.toolkits.timeseries as ts
+import operational_analysis.toolkits.unit_conversion as un
+import operational_analysis.toolkits.met_data_processing as met
+from operational_analysis.toolkits import filters
+
+from operational_analysis import logged_method_call
+from operational_analysis import logging
+
+logger = logging.getLogger(__name__)
+
+class Project_Engie(PlantData):
+    """This class loads data for the ENGIE La Haute Borne site into a PlantData
+    object"""
+
+    def __init__(self, path="data/la_haute_borne", name="Engie",
+                engine="pandas", toolkit=["pruf_plant_analysis"]):
+
+        super(Project_Engie, self).__init__(path, name, engine, toolkit)
+
+    def prepare(self):
+        """
+        Do all loading and preparation of the data for this plant.
+        """     
+        # Set time frequencies of data in minutes
+        self._meter_freq = '10T'  # Daily meter data
+        self._curtail_freq = '10T'  # Daily curtailment data
+        self._scada_freq = '10T' # 10-min 
+        
+        # Load meta data
+        self._lat_lon = (48.452, 5.588)
+        self._plant_capacity = 8.2 # MW
+        self._num_turbines = 4
+        self._turbine_capacity = 2.05 # MW
+        
+        ###################
+        # SCADA DATA #
+        ###################
+        logger.info("Loading SCADA data")
+        self._scada.load(self._path, "la-haute-borne-data-2014-2015", "csv")  # Load Scada data
+        logger.info("SCADA data loaded")
+        
+        logger.info("Timestamp QC and conversion to UTC")
+        # Get 'time' field in datetime format. Local time zone information is
+        # encoded, so convert to UTC
+        self._scada.df['time'] = pd.to_datetime(self._scada.df['Date_time'],utc=True)
+        
+        # Remove duplicated timestamps and turbine id
+        self._scada.df = self._scada.df.drop_duplicates(subset=['time','Wind_turbine_name'],keep='first')
+        
+        # Set time as index
+        self._scada.df.set_index('time',inplace=True,drop=False) 
+        
+        logger.info("Correcting for out of range of temperature variables")
+        # Handle extrema values for temperature. All other variables appear to
[…]
