### 2020-06-04 5fb9b84
saving partially refactored class
