+++ operational_analysis/methods/turbine_long_term_gross_energy.py
@@ -92,14 +92,9 @@ class TurbineLongTermGrossEnergy(object):
         self._model_results = {}
         self._turb_lt_gross = {}
         self._scada_daily_valid = pd.DataFrame()
-
-        # Dictionary to convert time interval indicator into minutes
-        self._time_conversion = {'10T': 10.,
-                                 '5T': 5.,
-                                 '1H': 60.}
         
         # Set number of 'valid' counts required when summing data to daily values
-        self._num_valid_daily = 60. / self._time_conversion[self._plant._scada_freq] * 24
+        self._num_valid_daily = 60. / (pd.to_timedelta(self._plant._scada_freq).seconds/60) * 24
         
         # Define relevant uncertainties, to be applied in Monte Carlo sampling
         self.uncertainty_scada = np.float64(uncertainty_scada)
@@ -132,13 +127,44 @@ class TurbineLongTermGrossEnergy(object):
         self.enable_plotting = enable_plotting 
         self.plot_dir = plot_dir
         
-        # Setup and run Monte Carlo
-        if self.UQ == True:
-            self.setup_TIE_monte_carlo_inputs()
-        self.run_TIE_monte_carlo()
+        self.setup_inputs()
+        
+        # Loop through number of simulations, store TIE results
+        for n in tqdm(np.arange(self.num_sim)):
+            
+            # MC-sampled parameter in this function!
+            logger.info("Filtering turbine data")
+            self.filter_turbine_data(n) # Filter turbine data
+            
+            if self.enable_plotting:
+                logger.info("Plotting filtered power curves")
+                self.plot_filtered_power_curves(self.plot_dir)
+            
+            # MC-sampled parameter in this function!
+            logger.info("Processing reanalysis data to daily averages")
+            self.setup_daily_reanalysis_data(n) # Setup daily reanalysis products
+            
+            # MC-sampled parameter in this function!
+            logger.info("Processing scada data to daily sums")
+            self.filter_sum_impute_scada(n) # Setup daily scada data
+            
+            logger.info("Setting up daily data for model fitting")
+            self.setup_model_dict() # Setup daily data to be fit using the GAM
+            
+            # MC-sampled parameter in this function!
+            logger.info("Fitting model data")
+            self.fit_model(n) # Fit daily turbine energy to atmospheric data
+            
+            logger.info("Applying fitting results to calculate long-term gross energy")
+            self.apply_model_to_lt(n) # Apply fitting result to long-term reanalysis data
+            
+            if self.enable_plotting:
+                logger.info("Plotting daily fitted power curves")
+                self.plot_daily_fitting_result(self.plot_dir) # Setup daily reanalysis products
         
         # Log the completion of the run
         logger.info("Run completed")
+        
 
     def sort_scada_by_turbine(self):
         """
@@ -177,12 +203,7 @@ class TurbineLongTermGrossEnergy(object):
         for t in self._turbs:
             turb_capac = dic[t].wtur_W_avg.max()
             
-            # Monte Carlo sample max_power_filter
-            if self.UQ == True:
-                self._max_power_filter = self._mc_max_power_filter[n]
-            elif self.UQ == False:
-                self._max_power_filter = np.mean(self.uncertainty_max_power_filter)
-            max_bin = self._max_power_filter * turb_capac # Set maximum range for using bin-filter
+            max_bin = self._input[n,"max_power_filter"] * turb_capac # Set maximum range for using bin-filter
             
             dic[t].dropna(subset = ['wmet_wdspd_avg', 'energy_kwh'], inplace = True) # Drop any data where scada wind speed or energy is NaN
             
[…]
