+++ operational_analysis/methods/turbine_long_term_gross_energy.py
@@ -132,6 +132,8 @@ class TurbineLongTermGrossEnergy(object):
         # Loop through number of simulations, store TIE results
         for n in tqdm(np.arange(self.num_sim)):
             
+            self._run = self._inputs[n]
+
             # MC-sampled parameter in this function!
             logger.info("Filtering turbine data")
             self.filter_turbine_data(n) # Filter turbine data
@@ -164,7 +166,42 @@ class TurbineLongTermGrossEnergy(object):
         
         # Log the completion of the run
         logger.info("Run completed")
-        
+    
+    @logged_method_call
+    def setup_inputs(self):
+        """
+        Create and populate the data frame defining the simulation parameters.
+        This data frame is stored as self._inputs
+
+        Args:
+            (None)
+            
+        Returns:
+            (None)
+        """
+        if self.UQ == True:
+            reanal_list = list(np.repeat(self._reanal, self.num_sim))  # Create extra long list of renanalysis product names to sample from
+            inputs = {
+                "reanalysis_product": np.asarray(random.sample(reanal_list, self.num_sim)),
+                "scada_data_fraction": np.random.normal(1, self.uncertainty_scada, self.num_sim),
+                "wind_bin_thresh": np.random.randint(self.uncertainty_wind_bin_thresh[0]*100, self.uncertainty_wind_bin_thresh[1]*100, self.num_sim) / 100.,
+                "max_power_filter": np.random.randint(self.uncertainty_max_power_filter[0]*100, self.uncertainty_max_power_filter[1]*100, self.num_sim) / 100.,
+                "correction_threshold": np.random.randint(self.uncertainty_correction_threshold[0]*100, self.uncertainty_correction_threshold[1]*100, self.num_sim) / 100.,
+            }
+            self._plant_gross = np.empty([self.num_sim,1]) 
+
+        if self.UQ == False:
+            inputs = {
+                "reanalysis_product": self._reanal,
+                "scada_data_fraction": 1,
+                "wind_bin_thresh": 2,
+                "max_power_filter": 0.85,
+                "correction_threshold": 0.90,
+            }
+            self._plant_gross = np.empty([len(self._reanal),1]) 
+
+        self._inputs = pd.DataFrame(inputs)
+
 
     def sort_scada_by_turbine(self):
         """
@@ -241,83 +278,8 @@ class TurbineLongTermGrossEnergy(object):
                        
             # Set negative turbine data to zero
             dic[t].loc[dic[t]['flag_neg'], 'wtur_W_avg'] = 0
-
-    def plot_filtered_power_curves(self, save_folder, output_to_terminal = False):
-        """
-        Plot the raw and flagged power curve data and save to file.
-        
-        Args:
-            save_folder('obj':'str'): The pathname to where figure files should be saved
-            output_to_terminal('obj':'boolean'): Indicate whether or not to output figures to terminal
-            
-        Returns:
-            (None)
-        """
-        import matplotlib.pyplot as plt
-        dic = self._scada_dict
-        
-        # Loop through turbines
-        for t in self._turbs:
-            filt_df = dic[t].loc[dic[t]['flag_final']] # Filter only for valid data
-            
-            plt.figure(figsize = (6,5))
-            plt.scatter(dic[t].wmet_wdspd_avg, dic[t].wtur_W_avg, s=1, label = 'Raw') # Plot all data
-            plt.scatter(filt_df['wmet_wdspd_avg'], filt_df['wtur_W_avg'], s=1, label = 'Flagged') # Plot flagged data
-            plt.xlim(0, 30)
-            plt.xlabel('Wind speed (m/s)')
-            plt.ylabel('Power (W)')
-            plt.title('Filtered power curve for Turbine %s' %t)
-            plt.legend(loc = 'lower right')
[…]
