### 2020-06-10 f2eb7e2
removed logged method call decorator
