### 2020-06-11 db0d497
tests now pass for turbine long term gross energy. UQ ifelse has been factored out of the class. I split the test case into two test cases for faster testing of the UQ=False case
