### 2020-06-11 b6df2c7
added simple memo to setup_daily_reanalysis_data
