### 2020-06-12 07d964b
removing requirement that gross energy must be non-negative when computing gross energy
