### 2020-06-12 d00a834
Re-factor electricl_losses and address Jordan's inline comments
