+++ operational_analysis/methods/turbine_long_term_gross_energy.py
@@ -47,16 +47,17 @@ class TurbineLongTermGrossEnergy(object):
     """
 
     @logged_method_call
-    def __init__(self, plant, UQ = True, num_sim = 100, uncertainty_scada=0.005,
-                  uncertainty_wind_bin_thresh=(1, 3), uncertainty_max_power_filter=(0.8, 0.9), 
-                  uncertainty_correction_threshold=(0.85, 0.95)):
+    def __init__(self, plant, UQ = False, num_sim = 2000, reanal_subset = ['erai', 'ncep2', 'merra2'],
+                 uncertainty_scada=0.005, uncertainty_wind_bin_thresh=(1, 3), 
+                 uncertainty_max_power_filter=(0.8, 0.9), uncertainty_correction_threshold=(0.85, 0.95)):
     
         """
         Initialize turbine long-term gross energy analysis with data and parameters.
         Args:
          plant(:obj:`PlantData object`): PlantData object from which TurbineLongTermGrossEnergy should draw data.
          UQ:(:obj:`bool`): choice whether to perform ('Y') or not ('N') uncertainty quantification
-         num_sim:(:obj:`int`): number of Monte Carlo simulations
+         num_sim:(:obj:`int`): number of Monte Carlo simulations. Please note that this script is somewhat computationally heavy so the default num_sim value has been adjusted accordingly.
+         reanal_subset(:obj:`list`): Which reanalysis products to use for long-term correction
          uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
          uncertainty_max_power_filter(:obj:`tuple`): Maximum power threshold (fraction) to which the bin filter 
                                             should be applied (default 0.85)
@@ -70,6 +71,11 @@ class TurbineLongTermGrossEnergy(object):
         if UQ == True:
             logger.info("Note: uncertainty quantification will be performed in the calculation")
             self.num_sim = num_sim
+            # Define relevant uncertainties, to be applied in Monte Carlo sampling
+            self.uncertainty_scada = uncertainty_scada
+            self.uncertainty_wind_bin_thresh = np.array(uncertainty_wind_bin_thresh, dtype=np.float64)
+            self.uncertainty_max_power_filter = np.array(uncertainty_max_power_filter, dtype=np.float64)
+            self.uncertainty_correction_threshold = np.array(uncertainty_correction_threshold, dtype=np.float64)
         elif UQ == False:    
             logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
             self.num_sim = None
@@ -79,6 +85,7 @@ class TurbineLongTermGrossEnergy(object):
         
         self._plant = plant  # Set plant as attribute of analysis object
         self._turbs = self._plant._scada.df['id'].unique() # Store turbine names
+        self._reanal = reanal_subset # Reanalysis data to consider in fitting
         
         # Get start and end of POR days in SCADA
         self._por_start = format(plant._scada.df.index.min(), '%Y-%m-%d')
@@ -96,25 +103,18 @@ class TurbineLongTermGrossEnergy(object):
         # Set number of 'valid' counts required when summing data to daily values
         self._num_valid_daily = 60. / (pd.to_timedelta(self._plant._scada_freq).seconds/60) * 24
         
-        # Define relevant uncertainties, to be applied in Monte Carlo sampling
-        self.uncertainty_scada = np.float64(uncertainty_scada)
-        self.uncertainty_wind_bin_thresh = np.array(uncertainty_wind_bin_thresh, dtype=np.float64)
-        self.uncertainty_max_power_filter = np.array(uncertainty_max_power_filter, dtype=np.float64)
-        self.uncertainty_correction_threshold = np.array(uncertainty_correction_threshold, dtype=np.float64)
-        
         # Initially sort the different turbine data into dictionary entries
[…]

[…]
