### 2020-06-12 b891f4b
updating qc notebook
