+++ operational_analysis/methods/quality_check_automation.py
@@ -20,7 +20,9 @@ class WindToolKitQualityControlDiagnosticSuite(object):
     the user can make informed decisions about how to handle the data."""
 
     @logged_method_call
-    def __init__(self,df, ws_field='wmet_wdspd_avg', power_field= 'wtur_W_avg', time_field= 'datetime', id_field= None, freq = '10T', lat_lon= (0,0), dst_subset = 'American'):
+    def __init__(self,df, ws_field='wmet_wdspd_avg', power_field= 'wtur_W_avg', time_field= 'datetime',
+                 id_field= None, freq = '10T', lat_lon= (0,0), dst_subset = 'American',
+                 check_tz = False):
         """
         Initialize QCAuto object with data and parameters.
 
@@ -33,9 +35,10 @@ class WindToolKitQualityControlDiagnosticSuite(object):
          freq(:obj: 'String'): String representation of the resolution for the time field to df
          lat_lon(:obj: 'tuple'): latitude and longitude of farm represented as a tuple
          dst_subset(:obj: 'String'): Set of Daylight Savings Time transitions to use (currently American or France)
+         check_tz(:obj: 'boolean'): Boolean on whether to use WIND Toolkit data to assess timezone of data
         """
         logger.info("Initializing QC_Automation Object")
-        
+
         self._df = df
         self._ws = ws_field
         self._w = power_field
@@ -44,6 +47,7 @@ class WindToolKitQualityControlDiagnosticSuite(object):
         self._freq = freq
         self._lat_lon = lat_lon
         self._dst_subset = dst_subset
+        self._check_tz = check_tz
 
         if self._id==None:
             self._id = 'ID'
@@ -53,10 +57,10 @@ class WindToolKitQualityControlDiagnosticSuite(object):
     def run(self):
         """
         Run the QC analysis functions in order by calling this function.
-        
+
         Args:
             (None)
-            
+
         Returns:
             (None)
         """
@@ -67,14 +71,17 @@ class WindToolKitQualityControlDiagnosticSuite(object):
         self.gap_time_identification()
         logger.info('Grabbing DST Transition Times')
         self.create_dst_df()
-        logger.info("Evaluating timezone deviation from UTC")
-        self.ws_diurnal_prep()
-        self.corr_df_calc()
+
+        if self._check_tz:
+            logger.info("Evaluating timezone deviation from UTC")
+            self.ws_diurnal_prep()
+            self.corr_df_calc()
+
         logger.info("Isolating Extrema Values")
         self.max_min()
         logger.info("QC Diagnostic Complete")
 
-        
+
     def dup_time_identification(self):
         """
         This function identifies any time duplications in the dataset.
@@ -102,7 +109,7 @@ class WindToolKitQualityControlDiagnosticSuite(object):
     def indicesForCoord(self,f):
 
         """
-        This function finds the nearest x/y indices for a given lat/lon. 
+        This function finds the nearest x/y indices for a given lat/lon.
         Rather than fetching the entire coordinates database, which is 500+ MB, this
         uses the Proj4 library to find a nearby point and then converts to x/y indices.
         This function relies on the Wind Toolkit HSDS API.
@@ -112,9 +119,9 @@ class WindToolKitQualityControlDiagnosticSuite(object):
         """
 
         dset_coords = f['coordinates']
-        projstring = """+proj=lcc +lat_1=30 +lat_2=60 
[…]
