### 2020-06-12 fe1dc0c
turb, elec, gap notebooks
