+++ examples/project_ENGIE.py
@@ -22,7 +22,7 @@ steps taken to correct the raw data for use in the PRUF OA code.
 - 10-minute availability and curtailment data in kwh
 - Generated by estimating availability and curtailment from SCADA data
 - Below-normal production classified as curtailment if present at all turbines.
-  Otherwise, classified as availability loss  
+  Otherwise, classified as availability loss
 - No need for timezone correction
 
 4. Reanalysis products
@@ -56,40 +56,45 @@ class Project_Engie(PlantData):
     def prepare(self):
         """
         Do all loading and preparation of the data for this plant.
-        """     
+        """
         # Set time frequencies of data in minutes
         self._meter_freq = '10T'  # Daily meter data
         self._curtail_freq = '10T'  # Daily curtailment data
-        self._scada_freq = '10T' # 10-min 
-        
+        self._scada_freq = '10T' # 10-min
+
         # Load meta data
         self._lat_lon = (48.452, 5.588)
         self._plant_capacity = 8.2 # MW
         self._num_turbines = 4
         self._turbine_capacity = 2.05 # MW
-        
+
         ###################
         # SCADA DATA #
         ###################
         logger.info("Loading SCADA data")
         self._scada.load(self._path, "la-haute-borne-data-2014-2015", "csv")  # Load Scada data
         logger.info("SCADA data loaded")
-        
+
         logger.info("Timestamp QC and conversion to UTC")
         # Get 'time' field in datetime format. Local time zone information is
         # encoded, so convert to UTC
-        self._scada.df['time'] = pd.to_datetime(self._scada.df['Date_time'],utc=True).dt.tz_localize(None)
-        
+
+        #self._scada.df['time'] = pd.to_datetime(self._scada.df['Date_time'],utc=True).dt.tz_localize(None)
+        date = [s[0:10] for s in self._scada.df['Date_time']]
+        time = [s[11:19] for s in self._scada.df['Date_time']]
+        datetime = [date[s] + ' ' + time[s] for s in np.arange(len(date))]
+        self._scada.df['time'] = pd.to_datetime(datetime, format = "%Y-%m-%d %H:%M:%S")
+
         # Remove duplicated timestamps and turbine id
         self._scada.df = self._scada.df.drop_duplicates(subset=['time','Wind_turbine_name'],keep='first')
-        
+
         # Set time as index
-        self._scada.df.set_index('time',inplace=True,drop=False) 
-        
+        self._scada.df.set_index('time',inplace=True,drop=False)
+
         logger.info("Correcting for out of range of temperature variables")
         # Handle extrema values for temperature. All other variables appear to
         # be reasonable.
-        self._scada.df = self._scada.df[(self._scada.df["Ot_avg"]>=-15.0) & (self._scada.df["Ot_avg"]<=45.0)]         
+        self._scada.df = self._scada.df[(self._scada.df["Ot_avg"]>=-15.0) & (self._scada.df["Ot_avg"]<=45.0)]
 
         logger.info("Flagging unresponsive sensors")
         # Due to data discretization, there appear to be a lot of repeating
@@ -103,7 +108,7 @@ class Project_Engie(PlantData):
             temp_flag = filters.unresponsive_flag(self._scada.df.loc[self._scada.df.Wind_turbine_name == id,'Ot_avg'], 20)
             self._scada.df.loc[(self._scada.df.Wind_turbine_name == id) \
                 & (temp_flag),'Ot_avg'] = np.nan
-        
+
         # Put power in watts
         self._scada.df["Power_W"] = self._scada.df["P_avg"] * 1000
 
@@ -111,22 +116,22 @@ class Project_Engie(PlantData):
[…]
