+++ CHANGELOG.md
@@ -1,0 +1,27 @@
+# Changelog
+All notable changes to this project will be documented in this file. If you make a notable change to the project, please add a line describing the change to the "unreleased" section. The maintainers will make an effort to keep the [Github Releases](https://github.com/NREL/OpenOA/releases) page up to date with this changelog. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [Unreleased]
+### Added
+- EYA Gap Analysis
+- Uncertainty quantification
+- Engie example data, and complete update of example notebooks
+
+## [1.1] - 2019-01-29
+- Python3 Support
+- Addition of reanalysis schemas to the Sphinx documentation
+- Easy import of EIA data using new module: Metadata_Fetch
+- Updated contributing.md document
+- Quality checks for reanalysis data
+- Improved installation instructions
+- Integration tests are now performed in CI
+- Performed PEP8 linting
+
+## [1.0] - 2018-12-06
+- Complete refactor of many analysis and toolkit modules.
+- Timeseries Table is now an integrated component, no sparkplug-datastructures dependency
+- Plant Level AEP method w/ Monte Carlo
+- Turbine / Scada level toolkits: Filtering, Imputing, Met, Pandas Plotting, Timeseries, Unit Conversion
+- Most toolkits and all methods are fully documented in Sphinx.
+- Two example notebooks: Operational AEP Analysis and Turbine Analysis
+- All toolkits except for Pandas Plotting have > 80% test coverage.
