+++ test/int_pruf_plant_analysis.py
@@ -1,42 +1,117 @@
 import unittest
-
 import numpy as np
 import pandas as pd
 from numpy import testing as nptest
 
 from operational_analysis.methods import plant_analysis
-from examples.operational_AEP_analysis.project_EIA import Project_EIA
+from examples.project_ENGIE import Project_Engie
 
 
 class TestPandasPrufPlantAnalysis(unittest.TestCase):
+    
     def setUp(self):
         np.random.seed(42)
 
-        # Set up data to use for testing (EIA example plant)
-        self.project = Project_EIA('./examples/operational_AEP_analysis/data')
+        # Set up data to use for testing (ENGIE example plant)
+        self.project = Project_Engie('./examples/data')
         self.project.prepare()
-        self.analysis = plant_analysis.MonteCarloAEP(self.project)
 
-    def test_validate_data(self):
-        self.assertTrue(self.project.validate(), "Failed to validate PlantData from schema")
-
-
+    # Test inputs to the regression model, at monthly time resolution
     def test_plant_analysis(self):
-
+        
+        # ____________________________________________________________________
+        # Test inputs to the regression model, at monthly time resolution
+        self.analysis = plant_analysis.MonteCarloAEP(self.project, 
+                                                      reanal_products=['merra2', 'era5'],
+                                                      time_resolution = 'M',
+                                                      reg_temperature = 'Y', 
+                                                      reg_winddirection = 'Y')
         df = self.analysis._aggregate.df
 
         # Check the pre-processing functions
-        self.check_process_revenue_meter_energy(df)
-        self.check_process_loss_estimates(df)
-        self.check_process_reanalysis_data(df)
+        self.check_process_revenue_meter_energy_monthly(df)
+        self.check_process_loss_estimates_monthly(df)
+        self.check_process_reanalysis_data_monthly(df)
 
+        # ____________________________________________________________________
+        # Test inputs to the regression model, at daily time resolution
+        self.analysis = plant_analysis.MonteCarloAEP(self.project, 
+                                                      reanal_products=['merra2', 'era5'],
+                                                      time_resolution = 'D',
+                                                      reg_temperature = 'Y', 
+                                                      reg_winddirection = 'Y')
+        df = self.analysis._aggregate.df
+        # Check the pre-processing functions
+        self.check_process_revenue_meter_energy_daily(df)
+        self.check_process_loss_estimates_daily(df)
+        self.check_process_reanalysis_data_daily(df)
+        
+        # ____________________________________________________________________
+        # Test linear regression model, at monthly time resolution
+        self.analysis = plant_analysis.MonteCarloAEP(self.project, 
+                                                     reanal_products=['merra2', 'era5'],
+                                                     time_resolution = 'M',
+                                                     reg_model = 'lin',
+                                                     reg_temperature = 'N', 
+                                                     reg_winddirection = 'N')
         # Run Monte Carlo AEP analysis, confirm the results are consistent
-        self.analysis.run(num_sim=2000, reanal_subset=['ncep2', 'merra2', 'erai'])
-
[…]
