+++ test/int_pruf_plant_analysis.py
@@ -13,7 +13,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
         np.random.seed(42)
 
         # Set up data to use for testing (ENGIE example plant)
-        self.project = Project_Engie('./examples/data')
+        self.project = Project_Engie('../examples/data')
         self.project.prepare()
 
     # Test inputs to the regression model, at monthly time resolution
