+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 All notable changes to this project will be documented in this file. If you make a notable change to the project, please add a line describing the change to the "unreleased" section. The maintainers will make an effort to keep the [Github Releases](https://github.com/NREL/OpenOA/releases) page up to date with this changelog. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-- EYA Gap Analysis
+- Energy Yield Analysis (EYA) to Operational Assessment(OA) Gap Analysis method
 - Uncertainty quantification
-- Engie example data, and complete update of example notebooks
+- Implemented open source Engie example data, and complete update of example notebooks
 - Switch to standard BSD-3 Clause license
-- Automated quality control method
+- Automated quality control method to assist with data ingestion. Tools in this method include daylight savings time change detection and identification of the diurnal cycle.
 - Electrical losses method
 - User-facing CI pipeline using Github Actions
 - Automatically built documentation using ReadTheDocs
@@ -22,7 +22,7 @@ All notable changes to this project will be documented in this file. If you make
 - Performed PEP8 linting
 
 ## [1.0] - 2018-12-06
-- Complete refactor of many analysis and toolkit modules.
+- Refactor many analysis and toolkit modules to make them conform to a standard API (init, prepare, and run method).
 - Timeseries Table is now an integrated component, no sparkplug-datastructures dependency
 - Plant Level AEP method w/ Monte Carlo
 - Turbine / Scada level toolkits: Filtering, Imputing, Met, Pandas Plotting, Timeseries, Unit Conversion
