### 2020-06-16 b354ab9
adding la haute born unzip to ci pipeline
