### 2020-06-17 2352b5c
updated contributing document to reference the changelog. Fixed badge in readme to point to develop branch
