### 2020-06-17 181aa79
upped version to 1.1.2 and fixed up some metadata in setup.py to prepare for pypi packaking
