+++ operational_analysis/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1"
+__version__ = "1.1.2"
 
 import json
 import logging
