+++ .github/workflows/tags-to-pypi.yml
@@ -1,0 +1,36 @@
+name: Publish any new version tags (matching 'v*') to PyPi as a new release.
+
+on:
+    push:
+        tags:
+            - 'v*'
+
+jobs:
+    build-n-publish:
+        name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+        runs-on: ubuntu-18.04
+        steps:
+            - uses: actions/checkout@master
+            - name: Set up Python 3.7
+              uses: actions/setup-python@v1
+              with:
+                python-version: 3.7
+            - name: Install pep517
+            run: >-
+                python -m
+                pip install
+                pep517
+                --user
+            - name: Build a binary wheel and a source tarball
+            run: >-
+                python -m
+                pep517.build
+                --source
+                --binary
+                --out-dir dist/
+                .
+            - name: Publish package to PyPI
+            uses: pypa/gh-action-pypi-publish@master
+            with:
+                password: ${{ secrets.PYPI_PASSWORD }}
+                repository_url: https://test.pypi.org/legacy/
