+++ .github/workflows/tags-to-pypi.yml
@@ -16,19 +16,19 @@ jobs:
               with:
                 python-version: 3.7
             - name: Install pep517
-            run: >-
-                python -m
-                pip install
-                pep517
-                --user
+                run: >-
+                    python -m
+                    pip install
+                    pep517
+                    --user
             - name: Build a binary wheel and a source tarball
-            run: >-
-                python -m
-                pep517.build
-                --source
-                --binary
-                --out-dir dist/
-                .
+                run: >-
+                    python -m
+                    pep517.build
+                    --source
+                    --binary
+                    --out-dir dist/
+                    .
             - name: Publish package to PyPI
             uses: pypa/gh-action-pypi-publish@master
             with:
