+++ .github/workflows/tags-to-pypi.yml
@@ -16,13 +16,13 @@ jobs:
               with:
                 python-version: 3.7
             - name: Install pep517
-                run: >-
+              run: >-
                     python -m
                     pip install
                     pep517
                     --user
             - name: Build a binary wheel and a source tarball
-                run: >-
+              run: >-
                     python -m
                     pep517.build
                     --source
@@ -30,7 +30,7 @@ jobs:
                     --out-dir dist/
                     .
             - name: Publish package to PyPI
-            uses: pypa/gh-action-pypi-publish@master
-            with:
+              uses: pypa/gh-action-pypi-publish@master
+              with:
                 password: ${{ secrets.PYPI_PASSWORD }}
                 repository_url: https://test.pypi.org/legacy/
