### 2020-06-17 7ee60f2
updating script
