+++ .github/workflows/tags-to-pypi.yml
@@ -1,4 +1,4 @@
-name: Publish any new version tags (matching 'v*') to PyPi as a new release.
+name: PyPi Publisher
 
 on:
     push:
@@ -7,7 +7,7 @@ on:
 
 jobs:
     build-n-publish:
-        name: Build and publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+        name: Build and publish Python package to PyPi
         runs-on: ubuntu-18.04
         steps:
             - uses: actions/checkout@master
@@ -15,21 +15,11 @@ jobs:
               uses: actions/setup-python@v1
               with:
                 python-version: 3.7
-            - name: Install pep517
-              run: >-
-                    python -m
-                    pip install
-                    pep517
-                    --user
-            - name: Build a binary wheel and a source tarball
-              run: >-
-                    python -m
-                    pep517.build
-                    --source
-                    --binary
-                    --out-dir dist/
-                    .
-            - name: Publish package to PyPI
+            - name: Build binary tarball
+              run: python setup.py bdist_wheel
+            - name: Build source tarball
+              run: python setup.py sdist --formats=gztar,zip
+            - name: Publish artifacts to PyPI
               uses: pypa/gh-action-pypi-publish@master
               with:
                 password: ${{ secrets.PYPI_PASSWORD }}
