+++ .github/workflows/tags-to-pypi.yml
@@ -16,9 +16,10 @@ jobs:
               with:
                 python-version: 3.7
             - name: Build tarballs
-              run: pip install wheel
-              run: python setup.py bdist_wheel
-              run: python setup.py sdist --formats=gztar,zip
+              run: |
+                pip install wheel
+                python setup.py bdist_wheel
+                python setup.py sdist --formats=gztar,zip
             - name: Publish artifacts to PyPI
               uses: pypa/gh-action-pypi-publish@master
               with:
