### 2020-06-17 4e67bb1
minor version increase for ci
