### 2020-06-17 20658a2
adding old prefix to ci script for old data
