### 2020-06-17 0c11b74
removed old from example path
