### 2020-06-17 e598c85
updated contributing to reflect packaging process
