### 2020-06-17 de8b80e
Re-factored plant_analysis
