+++ operational_analysis/methods/plant_analysis.py
@@ -30,7 +30,6 @@ from operational_analysis import logging
 
 logger = logging.getLogger(__name__)
 
-
 class MonteCarloAEP(object):
     """
     A serial (Pandas-driven) implementation of the benchmark PRUF operational
@@ -52,8 +51,7 @@ class MonteCarloAEP(object):
 
     @logged_method_call
     def __init__(self, plant, reanal_products=["merra2", "ncep2", "erai"], uncertainty_meter=0.005, uncertainty_losses=0.05,
-                 uncertainty_windiness=(10, 20), uncertainty_wind_bin_thresh=(1, 3), 
-                 uncertainty_loss_max=(10, 20), uncertainty_max_power_filter=(0.8, 0.9), 
+                 uncertainty_windiness=(10, 20), uncertainty_loss_max=(10, 20),
                  uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', 
                  reg_temperature = 'N', reg_winddirection = 'N'):
         """
@@ -65,10 +63,7 @@ class MonteCarloAEP(object):
          uncertainty_meter(:obj:`float`): uncertainty on revenue meter data
          uncertainty_losses(:obj:`float`): uncertainty on long-term losses
          uncertainty_windiness(:obj:`float`): number of years to use for the windiness correction
-         uncertainty_wind_bin_thresh(:obj:`float`): The filter threshold for each bin (default is 2 m/s)
          uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
-         uncertainty_max_power_filter(:obj:`float`): Maximum power threshold (fraction) to which the bin filter 
-                                            should be applied (default 0.85)
          uncertainty_nan_energy(:obj:`float`): threshold to flag days/months based on NaNs
          time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
          reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr')
@@ -90,8 +85,6 @@ class MonteCarloAEP(object):
         self.uncertainty_losses = np.float64(uncertainty_losses)
         self.uncertainty_windiness = np.array(uncertainty_windiness, dtype=np.float64)
         self.uncertainty_loss_max = np.array(uncertainty_loss_max, dtype=np.float64)
-        self.uncertainty_max_power_filter = np.array(uncertainty_max_power_filter, dtype=np.float64)
-        self.uncertainty_wind_bin_thresh = np.array(uncertainty_wind_bin_thresh, dtype=np.float64)
         self.uncertainty_nan_energy = np.float64(uncertainty_nan_energy)
         
         # Check that selected time resolution is allowed
@@ -101,8 +94,9 @@ class MonteCarloAEP(object):
         
         self._resample_freq = {"M": 'MS', "D": 'D'}[self.time_resolution]
         self._hours_in_res = {"M": 366*24, "D": 1*24}[self.time_resolution]
+        self._calendar_samples = {"M": 12, "D": 365}[self.time_resolution]
         if self.time_resolution == 'M':
-            self.num_days_lt= (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
+            self.num_days_lt = (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
 
         # Check that choices for regression inputs are allowed
         if reg_temperature not in ['Y', 'N']:
@@ -131,10 +125,8 @@ class MonteCarloAEP(object):
         # Create a data frame to store monthly/daily reanalysis data over plant period of record
[…]
