### 2020-06-19 005a3d5
Change l5p to return 0 for input values of 0
Wrap the logistic 5 parameter power curve function in an np.where to replace all cases where x=0 with 0, avoiding the divide by zero error.
