### 2020-06-19 e7758ce
Whoops, it was actually supposed to be "d" for the case x=0.
