### 2020-06-19 2039c53
Speed up the test
