### 2020-06-19 2bbb81b
Added separate case for b<0 where d is selected in the case of x==0.
