### 2020-06-20 e1d0299
Clean up code a little
