### 2020-06-21 f5cf1b8
Minor changes on doc description of a function
