### 2020-06-21 3f7f252
Let user choose filtering parameters for UQ = False case
