+++ test/int_turbine_long_term_gross_energy.py
@@ -15,7 +15,11 @@ class TestLongTermGrossEnergy(unittest.TestCase):
         self.project = TurbineExampleProject('./examples/turbine_analysis/data')
         self.project.prepare()
         
-        self.analysis = TurbineLongTermGrossEnergy(self.project, UQ = False, reanal_subset = ['erai', 'merra2', 'ncep2'])
+        self.analysis = TurbineLongTermGrossEnergy(self.project, UQ = False, 
+                                                   reanal_subset = ['erai', 'merra2', 'ncep2'],
+                                                   uncertainty_max_power_filter = 0.85,
+                                                   uncertainty_wind_bin_thresh = 1,
+                                                   uncertainty_correction_threshold = 0.9)
         self.analysis.run(enable_plotting = False)
 
     def test_longterm_gross_energy_results(self):
