### 2020-06-22 e227c02
Make TIE and electrical losses backward compatible
