+++ test/int_turbine_long_term_gross_energy.py
@@ -15,12 +15,13 @@ class TestLongTermGrossEnergy(unittest.TestCase):
         self.project = TurbineExampleProject('./examples/turbine_analysis/data')
         self.project.prepare()
         
-        self.analysis = TurbineLongTermGrossEnergy(self.project, UQ = False, 
-                                                   reanal_subset = ['erai', 'merra2', 'ncep2'],
-                                                   uncertainty_max_power_filter = 0.85,
-                                                   uncertainty_wind_bin_thresh = 1,
-                                                   uncertainty_correction_threshold = 0.9)
-        self.analysis.run(enable_plotting = False)
+        self.analysis = TurbineLongTermGrossEnergy(self.project, UQ = False)
+        
+        self.analysis.run(reanal_subset = ['erai', 'merra2', 'ncep2'],
+                                                   max_power_filter = 0.85,
+                                                   wind_bin_thresh = 1,
+                                                   correction_threshold = 0.9,
+                                                   enable_plotting = False)
 
     def test_longterm_gross_energy_results(self):
 
@@ -41,8 +42,8 @@ class TestLongTermGrossEnergyUQ(unittest.TestCase):
         self.project = TurbineExampleProject('./examples/turbine_analysis/data')
         self.project.prepare()
 
-        self.analysis_uq = TurbineLongTermGrossEnergy(self.project, UQ = True, num_sim = 100, reanal_subset = ['erai', 'merra2', 'ncep2'])
-        self.analysis_uq.run(enable_plotting = False)
+        self.analysis_uq = TurbineLongTermGrossEnergy(self.project, UQ = True, num_sim = 100)
+        self.analysis_uq.run(enable_plotting = False, reanal_subset = ['erai', 'merra2', 'ncep2'])
 
     def test_longterm_gross_energy_results(self):
 
