+++ operational_analysis/methods/turbine_long_term_gross_energy.py
@@ -47,9 +47,7 @@ class TurbineLongTermGrossEnergy(object):
     """
 
     @logged_method_call
-    def __init__(self, plant, UQ = False, num_sim = 2000, reanal_subset = ['erai', 'ncep2', 'merra2'],
-                 uncertainty_scada=0.005, uncertainty_wind_bin_thresh=(1, 3), 
-                 uncertainty_max_power_filter=(0.8, 0.9), uncertainty_correction_threshold=(0.85, 0.95)):
+    def __init__(self, plant, UQ = False, num_sim = 2000):
     
         """
         Initialize turbine long-term gross energy analysis with data and parameters.
@@ -57,13 +55,6 @@ class TurbineLongTermGrossEnergy(object):
          plant(:obj:`PlantData object`): PlantData object from which TurbineLongTermGrossEnergy should draw data.
          UQ:(:obj:`bool`): choice whether to perform ('Y') or not ('N') uncertainty quantification
          num_sim:(:obj:`int`): number of Monte Carlo simulations. Please note that this script is somewhat computationally heavy so the default num_sim value has been adjusted accordingly.
-         reanal_subset(:obj:`list`): Which reanalysis products to use for long-term correction
-         uncertainty_scada(:obj:`float`): uncertainty imposed to scada data
-         uncertainty_max_power_filter(:obj:`tuple`): Maximum power threshold (fraction) to which the bin filter 
-                                            should be applied (default 0.85)
-         uncertainty_wind_bin_thresh(:obj:`tuple`): The filter threshold for each bin (default is 2 m/s)
-         uncertainty_correction_threshold(:obj:`tuple`): The threshold (fraction) above which daily scada energy data
-                                                should be corrected (default is 0.90)
         """
         logger.info("Initializing TurbineLongTermGrossEnergy Object")
         
@@ -71,8 +62,6 @@ class TurbineLongTermGrossEnergy(object):
         if UQ == True:
             logger.info("Note: uncertainty quantification will be performed in the calculation")
             self.num_sim = num_sim
-            # Define relevant uncertainties, to be applied in Monte Carlo sampling
-            self.uncertainty_scada = uncertainty_scada
         elif UQ == False:    
             logger.info("Note: uncertainty quantification will NOT be performed in the calculation")
             self.num_sim = None
@@ -82,13 +71,7 @@ class TurbineLongTermGrossEnergy(object):
         
         self._plant = plant  # Set plant as attribute of analysis object
         self._turbs = self._plant._scada.df['id'].unique() # Store turbine names
-        self._reanal = reanal_subset # Reanalysis data to consider in fitting
-        
-        # Define relevant uncertainties, to be applied in Monte Carlo sampling
-        self.uncertainty_wind_bin_thresh = np.array(uncertainty_wind_bin_thresh, dtype=np.float64)
-        self.uncertainty_max_power_filter = np.array(uncertainty_max_power_filter, dtype=np.float64)
-        self.uncertainty_correction_threshold = np.array(uncertainty_correction_threshold, dtype=np.float64)
-        
+
         # Get start and end of POR days in SCADA
         self._por_start = format(plant._scada.df.index.min(), '%Y-%m-%d')
         self._por_end = format(plant._scada.df.index.max(), '%Y-%m-%d')
@@ -111,14 +94,24 @@ class TurbineLongTermGrossEnergy(object):
         
 
     @logged_method_call
-    def run(self, enable_plotting = False,
-            plot_dir = None):
+    def run(self,reanal_subset = ['erai', 'ncep2', 'merra2'],
+                 uncertainty_scada=0.005, wind_bin_thresh=(1, 3), 
[…]
