### 2020-06-22 cd44bd7
Default datatype is now float
