### 2020-06-23 #2171 @gschivley
@arfon and @sjpfenninger COVID and kids at home threw my schedule for a loop but I'm prioritizing this now. Writing up notes on issues I've found with docs and tests right now. Do I include those here or open new issues in the OpenOA repo?
