### 2020-06-23 #2171 @arfon
> Writing up notes on issues I've found with docs and tests right now. Do I include those here or open new issues in the OpenOA repo?

Please open issues in the OpenOA repo but mention this thread (https://github.com/openjournals/joss-reviews/issues/2171) in the comment which will cross link them for visibility.
