### 2020-06-23 NREL/OpenOA#78 @gschivley
Difference in README and docs testing instructions
Hi all - sorry for the very delayed review. I'm trying to work through this quickly now.

When trying to run unit/integration tests for OpenOA I noticed that the README and docs have different instructions.

- The README uses `python setup.py test`, the docs recommend `pytest -o python_files=test/*.py --cov=operational_analysis`

Using `python setup.py test` gave a depreciation warning
```
WARNING: Testing via this command is deprecated and will be removed in a future version. Users looking for a generic test entry point independent of test runner are encouraged to use tox.
```
and when I tried `pytest -o python_files=test/*.py --cov=operational_analysis` it did not recognize the `python_files` flag. I haven't used this before but it seems that this flag is supposed to be used in a [configuration file](https://docs.pytest.org/en/latest/example/pythoncollection.html#changing-naming-conventions)?

- The README does not include the final step when [extracting example data](https://openoa.readthedocs.io/en/latest/install.html#extracting-example-data), which leads to errors.
```
cp examples/turbine_analysis/data/example_20180829/scada_10min_4cols.csv examples/turbine_analysis/data/scada_10min_4cols.csv
```

Part of review (openjournals/joss-reviews#2171)
