### 2020-06-23 NREL/OpenOA#79 @gschivley
Depreciated pandas attribute .as_matrix
When running tests I found 3 spots where Pandas `as_matrix` is used. This attribute was [depreciated in 0.25](https://pandas.pydata.org/pandas-docs/version/0.25.1/reference/api/pandas.DataFrame.as_matrix.html) and I'm guessing is removed in 1.0. Since the [installation instructions](https://github.com/NREL/OpenOA#installation) don't use the `requirements.txt` file, the most recent version of pandas is installed.

```

========================================================================================= short test summary info ==========================================================================================
FAILED test/int_pruf_plant_analysis.py::TestPandasPrufPlantAnalysis::test_plant_analysis - AttributeError: 'Series' object has no attribute 'as_matrix'
FAILED test/test_imputing_toolkit.py::SimpleFilters::test_impute_data - AttributeError: 'Series' object has no attribute 'as_matrix'
FAILED test/test_met_data_processing_toolkit.py::SimpleMetProcessing::test_compute_shear - AttributeError: 'DataFrame' object has no attribute 'as_matrix'
================================================================================ 3 failed, 36 passed, 5 warnings in 43.87s =================================================================================
```

Part of review (https://github.com/openjournals/joss-reviews/issues/2171)
