### 2020-06-23 0186ae0
Incorporating Jordan's comments
