+++ operational_analysis/methods/plant_analysis.py
@@ -11,8 +11,6 @@
 import random
 import numpy as np
 import pandas as pd
-import matplotlib
-matplotlib.use('Agg')
 import statsmodels.api as sm
 from tqdm import tqdm
 from sklearn.linear_model import LinearRegression
@@ -52,7 +50,7 @@ class MonteCarloAEP(object):
     @logged_method_call
     def __init__(self, plant, reanal_products=["merra2", "ncep2", "erai"], uncertainty_meter=0.005, uncertainty_losses=0.05,
                  uncertainty_windiness=(10, 20), uncertainty_loss_max=(10, 20),
-                 uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', 
+                 uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', ml_setup_kwargs={},
                  reg_temperature = 'N', reg_winddirection = 'N'):
         """
         Initialize APE_MC analysis with data and parameters.
@@ -66,7 +64,8 @@ class MonteCarloAEP(object):
          uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
          uncertainty_nan_energy(:obj:`float`): threshold to flag days/months based on NaNs
          time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
-         reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr')
+         reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr'). At monthly time resolution only linear regression is allowed because of the reduced number of data points.
+         ml_setup_kwargs(:obj:`kwargs`): keyword arguments to MachineLearningSetup class
          reg_temperature(:obj:`string`): whether to include temperature ('Y') or not ('N') as regression input
          reg_winddirection(:obj:`string`): whether to include wind direction ('Y') or not ('N') as regression input
 
@@ -91,12 +90,10 @@ class MonteCarloAEP(object):
         if time_resolution not in ['M','D']:
             raise ValueError("time_res has to either be M (monthly, default) or D (daily)")
         self.time_resolution = time_resolution
-        
         self._resample_freq = {"M": 'MS', "D": 'D'}[self.time_resolution]
         self._hours_in_res = {"M": 366*24, "D": 1*24}[self.time_resolution]
         self._calendar_samples = {"M": 12, "D": 365}[self.time_resolution]
-        if self.time_resolution == 'M':
-            self.num_days_lt = (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
+        self.num_days_lt = (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
 
         # Check that choices for regression inputs are allowed
         if reg_temperature not in ['Y', 'N']:
@@ -110,8 +107,9 @@ class MonteCarloAEP(object):
         if reg_model not in ['lin', 'gbm','etr','gam']:
             raise ValueError("reg_model has to either be lin (Linear regression, default), gbm (Gradient boosting model), etr (Extra trees regressor) or gam (Generalized additive model)")
         self.reg_model = reg_model
+        self.ml_setup_kwargs = ml_setup_kwargs
         
-        # Monthly data can only use robus linear regression
+        # Monthly data can only use robust linear regression because of limited number of data
         if (time_resolution == 'M') & (reg_model != 'lin'):
             raise ValueError("For monthly time resolution, only linear regression is allowed!")
 
@@ -142,9 +140,9 @@ class MonteCarloAEP(object):
         self.num_sim = num_sim
         
         if reanal_subset is None:
[…]
