### 2020-06-23 #2171 @jordanperr
Hi all - Both of these issues have been fixed already in the develop branch. However, we are in the middle of a push for a new release, so develop is now incoherent in other ways. I don't think it's worth it to back-port these PRs into the joss-paper branch yet. Instead, I will perform the merge after the release (probably ~2-3 weeks out).

Is this acceptable to the reviewers?
