### 2020-06-23 NREL/OpenOA#78 @jordanperr
As mentioned in the Review ticket, this has been fixed already in develop.
