### 2020-06-23 01ce187
Fixing class for tests...
