+++ operational_analysis/methods/electrical_losses.py
@@ -79,12 +79,12 @@ class ElectricalLosses(object):
         Returns:
             (None)
         """
-        
+        # Define uncertainties
+        self.uncertainty_correction_thresh = np.array(uncertainty_correction_thresh, dtype=np.float64)  
         if self.UQ == True:
-            # Define uncertainties
             self.uncertainty_meter = uncertainty_meter
             self.uncertainty_scada = uncertainty_scada
-            self.uncertainty_correction_thresh = np.array(uncertainty_correction_thresh, dtype=np.float64)  
+        
                
         # Process SCADA data to daily sums
         self.process_scada()
