### 2020-06-23 a52bce6
Update electrical_losses.py
