### 2020-06-23 675d297
Update int_electrical_losses.py
