### 2020-06-23 #2171 @gschivley
@jordanperr I see that `as_matrix` is used in at least [one part of the production code](https://github.com/NREL/OpenOA/blob/cc2e56823d0c9cf46821d277efb10a799fc851a6/operational_analysis/toolkits/met_data_processing.py#L190). I don't have a clear sense of how anticipated updates after review are usually handled within JOSS, but it seems like that should be fixed. Alternatively, the docs could give users a way to specify an earlier version of Pandas where `as_matrix` works.
