### 2020-06-23 02ac169
Added logo to readme doc
