### 2020-06-24 9f88866
Check input types
