+++ operational_analysis/methods/plant_analysis.py
@@ -51,7 +51,7 @@ class MonteCarloAEP(object):
     def __init__(self, plant, reanal_products=["merra2", "ncep2", "erai"], uncertainty_meter=0.005, uncertainty_losses=0.05,
                  uncertainty_windiness=(10, 20), uncertainty_loss_max=(10, 20),
                  uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', ml_setup_kwargs={},
-                 reg_temperature = 'N', reg_winddirection = 'N'):
+                 reg_temperature = False, reg_winddirection = False):
         """
         Initialize APE_MC analysis with data and parameters.
 
@@ -66,8 +66,8 @@ class MonteCarloAEP(object):
          time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
          reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr'). At monthly time resolution only linear regression is allowed because of the reduced number of data points.
          ml_setup_kwargs(:obj:`kwargs`): keyword arguments to MachineLearningSetup class
-         reg_temperature(:obj:`string`): whether to include temperature ('Y') or not ('N') as regression input
-         reg_winddirection(:obj:`string`): whether to include wind direction ('Y') or not ('N') as regression input
+         reg_temperature(:obj:`bool`): whether to include temperature (True) or not (False) as regression input
+         reg_winddirection(:obj:`bool`): whether to include wind direction (True) or not (False) as regression input
 
         """
         logger.info("Initializing MonteCarloAEP Analysis Object")
@@ -96,10 +96,10 @@ class MonteCarloAEP(object):
         self.num_days_lt = (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
 
         # Check that choices for regression inputs are allowed
-        if reg_temperature not in ['Y', 'N']:
-            raise ValueError("reg_temperature has to either be Y (if temperature is considered in the regression), or N (if temperature is omitted")
-        if reg_winddirection not in ['Y', 'N']:
-            raise ValueError("reg_winddirection has to either be Y (if wind direction is considered in the regression), or N (if wind direction is omitted")
+        if reg_temperature not in [True, False]:
+            raise ValueError("reg_temperature has to either be True (if temperature is considered in the regression), or False (if temperature is omitted")
+        if reg_winddirection not in [True, False]:
+            raise ValueError("reg_winddirection has to either be True (if wind direction is considered in the regression), or False (if wind direction is omitted")
         self.reg_winddirection = reg_winddirection
         self.reg_temperature = reg_temperature
         
@@ -217,7 +217,6 @@ class MonteCarloAEP(object):
         """
         import matplotlib.pyplot as plt
         valid_aggregate = self._aggregate.df
-        project = self._plant
         plt.figure(figsize=(9, 9))
 
         # Loop through each reanalysis product and make a scatterplot of monthly wind speed vs plant energy
@@ -506,10 +505,10 @@ class MonteCarloAEP(object):
                                                                          rean_df['rho_kgm-3'])  # Density correct wind speeds
             self._reanalysis_aggregate[key] = rean_df.resample(self._resample_freq)['ws_dens_corr'].mean()  # .to_frame() # Get average wind speed by year-month
 
-            if self.reg_temperature == 'Y': # if temperature is considered as regression variable
+            if self.reg_temperature == True: # if temperature is considered as regression variable
                 self._reanalysis_aggregate[key + '_temp'] = pd.to_numeric(rean_df['temperature_K']).resample(self._resample_freq).mean()  # .to_frame() # Get average temperature by year-month
             
[…]
