+++ operational_analysis/methods/electrical_losses.py
@@ -80,13 +80,11 @@ class ElectricalLosses(object):
             (None)
         """
         # Define uncertainties and check types
-        if self.UQ == False:
-            if type(uncertainty_correction_thresh) != float:
-                logger.info("uncertainty_correction_thresh needs to be a float if UQ is False!")
+        expected_type = float if self.UQ == False else tuple
+        assert type(uncertainty_correction_thresh) == expected_type,  f"uncertainty_correction_thresh must be {expected_type} for UQ={self.UQ}"
+
         self.uncertainty_correction_thresh = np.array(uncertainty_correction_thresh, dtype=np.float64)  
         if self.UQ == True:
-            if type(uncertainty_correction_thresh) != tuple:
-                logger.info("uncertainty_correction_thresh needs to be a tuple if UQ is True!")
             self.uncertainty_meter = uncertainty_meter
             self.uncertainty_scada = uncertainty_scada
         
