### 2020-06-24 83bd74b
Update int_turbine_long_term_gross_energy.py
