### 2020-06-25 c526492
Eric's comments + one typo
