+++ operational_analysis/methods/plant_analysis.py
@@ -33,9 +33,7 @@ class MonteCarloAEP(object):
     A serial (Pandas-driven) implementation of the benchmark PRUF operational
     analysis implementation. This module collects standard processing and
     analysis methods for estimating plant level operational AEP and uncertainty.
-
     The preprocessing should run in this order:
-
         1. Process revenue meter energy - creates monthly/daily data frame, gets revenue meter on monthly/daily basis, and adds
            data flag
         2. Process loss estimates - add monthly/daily curtailment and availabilty losses to monthly/daily data frame
@@ -43,7 +41,6 @@ class MonteCarloAEP(object):
             from several reanalysis products to the monthly data frame
         4. Set up Monte Carlo - create the necessary Monte Carlo inputs to the OA process
         5. Run AEP Monte Carlo - run the OA process iteratively to get distribution of AEP results
-
     The end result is a distribution of AEP results which we use to assess expected AEP and associated uncertainty
     """
 
@@ -54,21 +51,19 @@ class MonteCarloAEP(object):
                  reg_temperature = False, reg_winddirection = False):
         """
         Initialize APE_MC analysis with data and parameters.
-
         Args:
          plant(:obj:`PlantData object`): PlantData object from which PlantAnalysis should draw data.
          reanal_products(obj:`list`) : List of reanalysis products to use for Monte Carlo sampling. Defaults to ["merra2", "ncep2", "erai"].
          uncertainty_meter(:obj:`float`): uncertainty on revenue meter data
          uncertainty_losses(:obj:`float`): uncertainty on long-term losses
-         uncertainty_windiness(:obj:`float`): number of years to use for the windiness correction
-         uncertainty_loss_max(:obj:`float`): threshold for the combined availabilty and curtailment monthly loss threshold
+         uncertainty_windiness(:obj:`tuple`): number of years to use for the windiness correction
+         uncertainty_loss_max(:obj:`tuple`): threshold for the combined availabilty and curtailment monthly loss threshold
          uncertainty_nan_energy(:obj:`float`): threshold to flag days/months based on NaNs
          time_resolution(:obj:`string`): whether to perform the AEP calculation at monthly ('M') or daily ('D') time resolution
          reg_model(:obj:`string`): which model to use for the regression ('lin' for linear, 'gam', 'gbm', 'etr'). At monthly time resolution only linear regression is allowed because of the reduced number of data points.
          ml_setup_kwargs(:obj:`kwargs`): keyword arguments to MachineLearningSetup class
          reg_temperature(:obj:`bool`): whether to include temperature (True) or not (False) as regression input
          reg_winddirection(:obj:`bool`): whether to include wind direction (True) or not (False) as regression input
-
         """
         logger.info("Initializing MonteCarloAEP Analysis Object")
 
@@ -91,7 +86,7 @@ class MonteCarloAEP(object):
             raise ValueError("time_res has to either be M (monthly, default) or D (daily)")
         self.time_resolution = time_resolution
         self._resample_freq = {"M": 'MS', "D": 'D'}[self.time_resolution]
-        self._hours_in_res = {"M": 366*24, "D": 1*24}[self.time_resolution]
+        self._hours_in_res = {"M": 30*24, "D": 1*24}[self.time_resolution]
         self._calendar_samples = {"M": 12, "D": 365}[self.time_resolution]
         self.num_days_lt = (31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
 
@@ -130,7 +125,6 @@ class MonteCarloAEP(object):
     def run(self, num_sim, reanal_subset=None):
         """
         Perform pre-processing of data into an internal representation for which the analysis can run more quickly.
-
         Args:
[…]
