### 2020-06-26 4c60278
Make TIE and electrical losses backward compatible + ENGIE tests (#77)
TIE and electrical losses should now be backward compatible, and they should now work with the Examples in PR #71 .
Also, following @ejsimley 's comments in PR #71 , filter and threshold values (max_power_filter, wind_bin_thresh, correction_threshold) are now not hardcoded anymore in the TIE class.

* Let user choose filtering parameters for UQ = False case
* Make TIE and electrical losses backward compatible
