+++ examples/turbine_analysis/engie_project.py
@@ -44,7 +44,7 @@ class TurbineEngieOpenData(PlantData):
                  self.scada.df.append(pd.read_json(url,orient='columns'), ignore_index=True)
         self.scada.rename_columns({"time": "date_time",
                                    "wtur_W_avg": "p_avg",
-                                   "wmet_wDir_avg": "wa_avg",
+                                   "wmet_wdir_avg": "wa_avg",
                                    "wmet_wdspd_avg": "ws_avg"})
         self.scada.df.set_index('time', inplace=True, drop=False)
         self.scada.df.drop(['date_time', 'p_avg', 'wa_avg', 'ws_avg'], axis=1, inplace=True)

[…]
