### 2020-06-27 f4a2239
First pass at Jordan's comments
