+++ operational_analysis/methods/plant_analysis.py
@@ -97,6 +97,13 @@ class MonteCarloAEP(object):
             raise ValueError("reg_winddirection has to either be True (if wind direction is considered in the regression), or False (if wind direction is omitted")
         self.reg_winddirection = reg_winddirection
         self.reg_temperature = reg_temperature
+
+        # Build list of regression variables
+        self._rean_vars = []
+        if self.reg_temperature:
+            self._rean_vars += ["temperature_K"]
+        if self.reg_winddirection:
+            self._rean_vars += ["u_ms", "v_ms"]
         
         # Check that selected regression model is allowed
         if reg_model not in ['lin', 'gbm','etr','gam']:
@@ -137,7 +144,7 @@ class MonteCarloAEP(object):
             self.reanal_subset = self._reanal_products
         else:
             self.reanal_subset = reanal_subset
-        
+                
         # Write parameters of run to the log file
         logged_self_params = ["uncertainty_meter", "uncertainty_losses", "uncertainty_loss_max", "uncertainty_windiness",
                               "uncertainty_nan_energy", "num_sim", "reanal_subset"]
@@ -482,16 +489,12 @@ class MonteCarloAEP(object):
             rean_df['ws_dens_corr'] = mt.air_density_adjusted_wind_speed(rean_df['windspeed_ms'],
                                                                          rean_df['rho_kgm-3'])  # Density correct wind speeds
             self._reanalysis_aggregate[key] = rean_df.resample(self._resample_freq)['ws_dens_corr'].mean()  # .to_frame() # Get average wind speed by year-month
-
-            if self.reg_temperature == True: # if temperature is considered as regression variable
-                self._reanalysis_aggregate[key + '_temp'] = pd.to_numeric(rean_df['temperature_K']).resample(self._resample_freq).mean()  # .to_frame() # Get average temperature by year-month
             
-            if self.reg_winddirection == True: # if wind direction is considered as regression variable
-                u_avg = pd.to_numeric(rean_df['u_ms']).resample(self._resample_freq).mean()  # .to_frame() # Get average u component by year-month
-                v_avg = pd.to_numeric(rean_df['v_ms']).resample(self._resample_freq).mean()  # .to_frame() # Get average v component by year-month
-                self._reanalysis_aggregate[key + '_u'] = u_avg 
-                self._reanalysis_aggregate[key + '_v'] = v_avg
-                self._reanalysis_aggregate[key + '_wd'] = 180-np.rad2deg(np.arctan2(-u_avg,v_avg)) # Calculate wind direction
+            namescol = [key + '_' + var for var in self._rean_vars]
+            self._reanalysis_aggregate[namescol] = rean_df[self._rean_vars].resample(self._resample_freq).mean()
+
+            if self.reg_winddirection: # if wind direction is considered as regression variable
+                self._reanalysis_aggregate[key + '_wd'] = np.pi-(np.arctan2(-self._reanalysis_aggregate[key + '_u_ms'],self._reanalysis_aggregate[key + '_v_ms'])) # Calculate wind direction
         
         self._aggregate.df = self._aggregate.df.join(
                 self._reanalysis_aggregate)  # Merge monthly reanalysis data to monthly energy data frame
@@ -624,13 +627,13 @@ class MonteCarloAEP(object):
         valid_data = df_sub.loc[df_sub.loc[:, 'flag_final'] == False, [reanal, 
                                                                'energy_gwh', 'availability_gwh',
                                                                'curtailment_gwh']]
-        if self.reg_winddirection == True:
[…]
