### 2020-06-30 c15acfb
cleaned up reanalysis function a little
