### 2020-07-01 6497f2a
adding prose to the examples
