### 2020-07-02 faaa129
Memoized approach for sample_long_term_reanalysis
