### 2020-07-02 d2a6c48
Delete plant_analysis.py
