### 2020-07-02 2d88e08
Memoized approach for sample_long_term_reanalysis
