### 2020-07-16 #2171 @gschivley
@sjpfenninger Sure. I still think the depreciated pandas `as_matrix` issue needs to be fixed before final approval. The default installation uses pandas 1.0.*, and there is at least one function that won't work with it.
