### 2020-07-17 #2171 @sjpfenninger
@gschivley thanks for pointing this out. I think critical requirements like pandas should be pinned in `setup.py`, e.g. with a maximum version number to use, so that the package doesn't break when pandas updates. @jordanperr It is perfectly fine to make fixes like this in response to the reviews here and to package them up in a new version once we've wrapped up the review. We can update the version on the paper on accept. Your joss-paper branch is primarily to get the paper itself ready to merge back to your master branch when we're ready to accept.
