### 2020-07-22 2e18614
adds the __init__ docstrings to the docs
