+++ sphinx/conf.py
@@ -183,3 +183,4 @@ texinfo_documents = [
 napoleon_google_docstring = True
 napoleon_use_param = False
 napoleon_use_ivar = False
+autoclass_content = 'both'
