### 2020-07-22 da8126b
tidies up docstrings
