+++ operational_analysis/methods/plant_analysis.py
@@ -33,14 +33,21 @@ class MonteCarloAEP(object):
     A serial (Pandas-driven) implementation of the benchmark PRUF operational
     analysis implementation. This module collects standard processing and
     analysis methods for estimating plant level operational AEP and uncertainty.
+    
     The preprocessing should run in this order:
+
         1. Process revenue meter energy - creates monthly/daily data frame, gets revenue meter on monthly/daily basis, and adds
            data flag
+
         2. Process loss estimates - add monthly/daily curtailment and availabilty losses to monthly/daily data frame
+        
         3. Process reanalysis data - add monthly/daily density-corrected wind speeds, temperature (if used) and wind direction (if used)
-            from several reanalysis products to the monthly data frame
+           from several reanalysis products to the monthly data frame
+        
         4. Set up Monte Carlo - create the necessary Monte Carlo inputs to the OA process
+        
         5. Run AEP Monte Carlo - run the OA process iteratively to get distribution of AEP results
+
     The end result is a distribution of AEP results which we use to assess expected AEP and associated uncertainty
     """
 
@@ -51,6 +58,7 @@ class MonteCarloAEP(object):
                  reg_temperature = False, reg_winddirection = False):
         """
         Initialize APE_MC analysis with data and parameters.
+        
         Args:
          plant(:obj:`PlantData object`): PlantData object from which PlantAnalysis should draw data.
          reanal_products(obj:`list`) : List of reanalysis products to use for Monte Carlo sampling. Defaults to ["merra2", "ncep2", "erai"].
@@ -132,11 +140,13 @@ class MonteCarloAEP(object):
     def run(self, num_sim, reanal_subset=None):
         """
         Perform pre-processing of data into an internal representation for which the analysis can run more quickly.
+        
         Args:
             reanal_subset(:obj:`list`): list of str data indicating which reanalysis products to use in OA
             num_sim(:obj:`int`): number of simulations to perform        
          
-        :return: None
+        Returns:
+            None
         """
         self.num_sim = num_sim
         
@@ -165,7 +175,9 @@ class MonteCarloAEP(object):
         """
         Make a plot of annual average wind speeds from reanalysis data to show general trends for each
         Highlight the period of record for plant data
-        :return: matplotlib.pyplot object
+        
+        Returns:
+            matplotlib.pyplot object
         """
         import matplotlib.pyplot as plt
         project = self._plant
@@ -209,9 +221,12 @@ class MonteCarloAEP(object):
     def plot_reanalysis_gross_energy_data(self, outlier_thres):
         """
         Make a plot of normalized 30-day gross energy vs wind speed for each reanalysis product, include R2 measure
-        :param outlier_thres (float): outlier threshold (typical range of 1 to 4) which adjusts outlier sensitivity
-        detection
-        :return: matplotlib.pyplot object
+        
+        Args:
+            outlier_thres (float): outlier threshold (typical range of 1 to 4) which adjusts outlier sensitivity detection
+        
+        Returns:
+            matplotlib.pyplot object
         """
         import matplotlib.pyplot as plt
         valid_aggregate = self._aggregate.df
@@ -251,7 +266,9 @@ class MonteCarloAEP(object):
     def plot_result_aep_distributions(self):
         """
         Plot a distribution of AEP values from the Monte-Carlo OA method
-        :return: matplotlib.pyplot object
+        
+        Returns:
+            matplotlib.pyplot object
         """
         import matplotlib.pyplot as plt
         fig = plt.figure(figsize=(14, 12))
@@ -284,9 +301,10 @@ class MonteCarloAEP(object):
     def plot_aep_boxplot(self, param, lab):
         """                                                                                                                                                                                        
         Plot box plots of AEP results sliced by a specified Monte Carlo parameter                                                                                                                  
+        
[…]
