### 2020-07-22 7c8ee48
update sphinx install section
