### 2020-07-23 31f98ff
remove old notebooks from documentation examples
