### 2020-07-29 7e267a9
deleting old examples
