### 2020-07-29 #2171 @jordanperr
Sounds good. Thanks for the comments so far! We are getting ready to release a new version of the software, and have decided to update the JOSS paper to reflect the new version. The latest code can be found in the develop branch: https://github.com/NREL/OpenOA/tree/develop I expect to merge this into the master branch within a couple weeks, and the paper will be updated shortly thereafter. Thanks for your patience!
