### 2020-07-30 NREL/OpenOA#79 @jordanperr
Closing this ticket, as_matrix has been removed in the develop branch.
