### 2020-07-30 NREL/OpenOA#78 @jordanperr
As mentioned, the unzipping issue has been fixed in develop.

I am unable to reproduce your issue with pytest. Can you please provide the version number you've used for pytest, and any relevant information about your development environment (mac vs windows) so that we can debug this together.

Thanks,
Jordan
