### 2020-07-30 ce920ff
Update plant_analysis.py
