### 2020-07-31 201d74c
adding nicolas text to paper
