+++ paper/paper.md
@@ -15,6 +15,9 @@ authors:
     affiliation: 1
   - name: Jason M. Fields
     affiliation: 1
+  - name: Nicola Bodini
+    orcid: 0000-0002-2550-9853
+    affiliation: 1
   - name: Lindy Williams
     affiliation: 1
   - name: Joseph C.Y. Lee
@@ -64,7 +67,7 @@ OpenOA was originally scoped to calculate the operational annual energy producti
 OpenOA has since expanded its scope to support additional types of analyses including turbine performance and electrical losses and to support data from wind plants outside the initial set of test projects.
 Released publicly in September 2018, the OpenOA repository contains numerous examples demonstrated in Jupyter notebooks along with publicly available data which can be used to run the built in unit and integration tests. The software has also been used in study on the impact of analyst choice in the outcome of OAs [@craig2018].
 
-# Importance of Operational Analysis
+## Importance of Operational Analysis
 Operational analysis uses collected data from wind farms to perform assessments ranging from the diagnosis of
 faults and underperformance, benchmarking of performance improvements (e.g., wind sector management,
 vortex generators), long-term estimates of AEP, and building
@@ -72,7 +75,9 @@ statistical or physics-based models for various applications (e.g.,, wake model 
 forecasting). Data sources include the wind farm revenue meter, turbine supervisory, control and data
 acquisition (SCADA) systems, on-site meteorological towers, and modeled atmospheric data (e.g., reanalysis data).
 
-## Long Term AEP Calculation
+# Built In Analysis Methods
+
+## Long Term Corrected AEP
 
 The AEP analysis implemented in OpenOA is based on a relatively standard approach within the wind resource assessment
 industry, where monthly gross energy for the wind plant (reported energy at the revenue meter corrected for availability
@@ -99,7 +104,41 @@ The details of this particular example are provided in a Jupyter notebook on the
 In this example, revenue meter and reanalysis data are processed using the data flagging, time series analysis,
 and plotting tools module to produce the graphs in the figure.*
 
-## Low Level Toolkits
+## Electrical Losses
+Electrical losses are calculated by comparing energy production measured by the turbine SCADA data with what is recorded at the revenue meter:
+
+Electrical Losses=1-(Revenue Meter Energy)/(Turbine SCADA Energy)
+
+For those projects that report data at daily or sub-daily time resolution, we calculate daily sums of energy production from the turbine SCADA and from the revenue meter data, only for days when all turbines within the wind farm are reporting at all time steps. We then calculate the average electrical losses, using the sum of the selected daily sums of revenue meter and turbine SCADA data over the period of record of each wind farm. When revenue meter data are only available at monthly time resolution (two projects in the set considered in our analysis), we filter out months in which there were less than an imposed percentage threshold (whose value is Monte Carlo sampled, as detailed below) of timestamps with all turbines within the wind farm running. The sum of filtered monthly revenue meter and turbine SCADA data over the period of record for each wind farm is then calculated and used to compute the average electrical loss.
+
+We quantify the uncertainty connected to the electrical loss calculation for each project by taking the standard deviation of the distribution of electrical losses created from a Monte Carlo simulation. For each project, the Monte Carlo simulation is run 10,000 times, and the following uncertainty components are included:
+
+- Revenue meter accuracy, which is incorporated into the Monte Carlo simulation by sampling monthly revenue meter data from a normal distribution centered on the reported value and with a standard deviation equal to 0.5% of the reported value. We consider 0.5% to be representative of what is typically assumed by wind energy consultants for revenue meter uncertainty.
+- SCADA data accuracy, also included in the Monte Carlo approach by sampling the data from a normal distribution centered on the recorded values and with a 0.5% uncertainty imposed.
[…]
