+++ operational_analysis/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.1.4"
+__version__ = "2.0.0"
 
 import json
 import logging
