### 2020-07-31 6b03735
adding tag instructions to contribting guidelines
