### 2020-07-31 bbc6952
Update plant_analysis.py
