### 2020-08-04 2f57687
Update plant_analysis.py
