+++ operational_analysis/methods/plant_analysis.py
@@ -778,11 +778,10 @@ class MonteCarloAEP(object):
             fitted_model = self.run_regression(n)
  
             # Get long-term regression inputs
-            [reg_inputs_lt_noIAV, reg_lt_stdev] = self.sample_long_term_reanalysis(self._run.num_years_windiness,
-                                                             self._run.reanalysis_product)
+            [reg_inputs_lt_noIAV, reg_lt_stdev] = self.sample_long_term_reanalysis()
 
             # Apply IAV to long-term regression inputs
-            reg_inputs_lt = self.apply_IAV(reg_inputs_lt_noIAV, reg_lt_stdev, self._run.reanalysis_product)
+            reg_inputs_lt = self.apply_IAV(reg_inputs_lt_noIAV, reg_lt_stdev)
             
             # Get long-term normalized gross energy by applying regression result to long-term monthly wind speeds
             inputs = np.array(reg_inputs_lt)
@@ -820,62 +819,60 @@ class MonteCarloAEP(object):
         return sim_results
 
     @logged_method_call
-    def sample_long_term_reanalysis(self, n, r):
+    def sample_long_term_reanalysis(self):
         """
         This function returns the windiness-corrected monthly wind speeds based on the Monte-Carlo generated sample of:
             1. The reanalysis product
             2. The number of years to use in the long-term correction
         Args:
-           n(:obj:`integer`): The number of years for the windiness correction
-           r(:obj:`string`): The reanalysis product used for the current Monte Carlo iteration
+            (None)
         Returns:
            :obj:`pandas.DataFrame`: the windiness-corrected or 'long-term' annualized monthly/daily wind speeds
            :obj:`pandas.DataFrame`: Long-term wind speed stdev, from which IAV will be derived
         """
         # Check if valid data has already been calculated and stored. If so, just return it
-        if (r, n) in self.long_term_sampling:
-            long_term_reg_inputs = self.long_term_sampling[(r, n)]
-            ws_aggregate_sd = self.long_term_sampling_sd[(r, n)]
+        if (self._run.reanalysis_product,self. _run.num_years_windiness) in self.long_term_sampling:
+            long_term_reg_inputs = self.long_term_sampling[(self._run.reanalysis_product,self. _run.num_years_windiness)]
+            ws_aggregate_sd = self.long_term_sampling_sd[(self._run.reanalysis_product,self. _run.num_years_windiness)]
             return [long_term_reg_inputs.copy(),ws_aggregate_sd.copy()]
         
         # Sample long-term wind speed values
-        ws_df = self._reanalysis_aggregate[r].to_frame().dropna()  # Drop NA values from monthly/daily reanalysis data series
-        ws_data = ws_df.tail(n * self._calendar_samples)  # Get last 'x' years of data from reanalysis product
-        ws_aggregate = self.groupby_time_res(ws_data)[r] # Get long-term annualized monthly/daily wind speeds
+        ws_df = self._reanalysis_aggregate[self._run.reanalysis_product].to_frame().dropna()  # Drop NA values from monthly/daily reanalysis data series
+        ws_data = ws_df.tail(self. _run.num_years_windiness * self._calendar_samples)  # Get last 'x' years of data from reanalysis product
+        ws_aggregate = self.groupby_time_res(ws_data)[self._run.reanalysis_product] # Get long-term annualized monthly/daily wind speeds
         ws_aggregate.reset_index(drop=True, inplace=True) # TODO: This rectifies a 0-indexing issue. Needs refactoring.
[…]
