### 2020-08-04 a70d437
Update plant_analysis.py
