+++ .github/workflows/tags-to-pypi.yml
@@ -24,4 +24,4 @@ jobs:
               uses: pypa/gh-action-pypi-publish@master
               with:
                 password: ${{ secrets.PYPI_PASSWORD }}
-                repository_url: https://test.pypi.org/legacy/
+                repository_url: https://pypi.org/legacy/
