### 2020-08-05 #2171 @sjpfenninger
I have invited a third reviewer so we can move forward here and hopefully wrap this up soon. Thanks for jumping in @ludee! @gschivley @ludee can you get your reviews done in the near future based on the https://github.com/NREL/OpenOA/tree/develop branch? @jordanperr can you clarify if you expect any major changes to the paper itself?
