### 2020-08-06 94e8406
Update electrical_losses.py
