### 2020-08-14 #2171 @jordanperr
@sjpfenninger OpenOA 2.0 is now released. We are now updating the paper. Please allow us some time to finish the update to the paper. I will ping you here when it's ready.
