### 2020-08-17 #2171 @jordanperr
Looks like it's been merged to me: https://github.com/NREL/OpenOA/blob/25960a95656197216a2dff8f4c990bde9d7b7146/operational_analysis/__init__.py#L1

As a reminder, the paper has not yet been updated.
