### 2020-08-17 0e4e9e1
moved paper v1 to paper_v1.md, started a new draft for v2
