+++ paper/paper.md
@@ -49,147 +49,62 @@ date: 20 December 2019
 bibliography: paper.bib
 ---
 
+<!--
+JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that explain the software functionality and domain of use to a non-specialist reader. We also require that authors explain the research applications of the software. The paper should be between 250-1000 words.
+
+Your paper should include:
+
+A list of the authors of the software and their affiliations, using the correct format (see the example below).
+A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
+A clear Statement of Need that illustrates the research purpose of the software.
+A list of key references, including to other software addressing related needs.
+Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
+Acknowledgement of any financial support.
+As this short list shows, JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of Need, Summary, Acknowledgements, and References sections. You can look at an example accepted paper. Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.
+
+
+Review Checklist:
+Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
+A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
+State of the field: Do the authors describe how this software compares to other commonly-used packages?
+Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
+References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
+
+-->
+
+This paper announces OpenOA version 2.0.0, which was released in August of 2020. Originally released to the public in 2018, OpenOA is implemented in the Python programming language [cite python] and provides a framework and set of convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate big data analysis of operational data sets from wind energy plants. With over 50 stars on github, a dozen contributors, and an active Github issues forum, OpenOA is a mature and active project that provides a high level interface to practical problems in this field.
+
+# Statement of Need
+
+Operational analyses are 
+
+OpenOA was created by and primarily developed by researchers at the National Renewable Energy Laboratory [nrel footnote] through the Performance, Risk Uncertainty Finance (PRUF) project. 
+
+OpenOA has also been used to calculate long-term operational AEP from over 470 wind farms in the US to assess correlation between uncertainty components [Bodini2020].
+
+Reproducible research is an important goal of OpenOA.
+
+OpenOA serves as a conduit to deliver state-of-the-art analysis methods between researchers and practitioners.
+
+We believe the reproducibility and efficiency of scientific research will increase if common, open source tools such as OpenOA are utilized.
+
+[Diagram of OA analysis]
+
 # Summary
 
-OpenOA is an open source Python package which implements operational analysis (OA) methods for wind energy plants.
-The goal of OpenOA is collaboration and methods sharing in a wind industry that has historically been very protective of methods and data [@McCann2018].
-Over time, we hope that OpenOA will become a reference implementation for OA methods.
+OpenOA V2 is released as a Python package and is freely available under a business-friendly, open-source, BSD license. OpenOA contains documentation, worked out examples in Jupyter notebooks, and an example dataset from Engie [cite engie].
 
-The development of OpenOA started internally at the National Renewable Energy Laboratory (NREL) to support the lab's
[…]
