### 2020-08-17 ddb2502
added some citations
