+++ paper/paper.bib
@@ -113,3 +113,73 @@ year = {2018}
   year={2020},
   abstract={Calculations of annual energy production (AEP) from a wind farm – whether based on preconstruction or operational data – are critical for wind farm financial transactions. The uncertainty in the AEP calculation is especially important in quantifying risk and is a key factor in determining financing terms. Standard industry practice assumes that different uncertainty categories within an AEP calculation are uncorrelated and can therefore be combined through a sum of squares approach. In this analysis, we assess the rigor of this assumption by performing operational AEP estimates for over 470 wind farms in the United States. We contrast the standard uncertainty assumption with a Monte Carlo approach to uncertainty quantification in which no assumptions of correlation between uncertainty categories are made. Results show that several uncertainty categories do, in fact, show weak to moderate correlations, namely: wind resource interannual variability and the windiness correction (positive correlation), wind resource interannual variability and regression (negative), and wind speed measurement uncertainty and regression (positive). The sources of these correlations are described and illustrated in detail in this paper, and the effect on the total AEP uncertainty calculation is investigated. Based on these results, we conclude that a Monte Carlo approach to AEP uncertainty quantification is more robust and accurate than the industry standard approach.}
 }
+
+
+
+@book{Rossum2009, 
+ author = {Van Rossum, Guido and Drake, Fred L.}, 
+ title = {Python 3 Reference Manual}, 
+ year = {2009}, 
+ isbn = {1441412697}, 
+ publisher = {CreateSpace}, 
+ address = {Scotts Valley, CA} 
+}
+
+@article{Pedregosa2011, 
+  title={Scikit-learn: Machine learning in Python}, 
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others}, 
+  journal={Journal of machine learning research}, 
+  volume={12}, 
+  number={Oct}, 
+  pages={2825--2830}, 
+  year={2011} 
+}
+
+@inproceedings{Mckinney2010, 
+  title={Data structures for statistical computing in python}, 
+  author={McKinney, Wes and others}, 
+  booktitle={Proceedings of the 9th Python in Science Conference}, 
+  volume={445}, 
+  pages={51--56}, 
+  year={2010}, 
+  organization={Austin, TX} 
+}
+
+@book{oliphant2006guide, 
+  title={A guide to NumPy}, 
+  author={Oliphant, Travis E}, 
+  volume={1}, 
+  year={2006}, 
+  publisher={Trelgol Publishing USA} 
+}
+
+
+@article{hunter2007matplotlib, 
+  title={Matplotlib: A 2D graphics environment}, 
+  author={Hunter, John D}, 
+  journal={Computing in science \& engineering}, 
+  volume={9}, 
+  number={3}, 
+  pages={90--95}, 
+  year={2007}, 
+  publisher={IEEE} 
+}
+
+
+@misc{EngieDataset,
+  title = {La Haute Borne Data (2013-2016)},
+  howpublished = {Engie Renewables. \url{https://opendata-renewables.engie.com/explore/dataset/d543716b-368d-4c53-8fb1-55addbe8d3ad/information}},
+  note = {Updated 9 October 2019}
+}
+
+
+@misc{osti_1478526,
+title = {Open OA, FKA: Wind Plant Performance Project (WP3) Benchmarking},
[…]
