+++ paper/paper.md
@@ -72,41 +72,39 @@ References: Is the list of references complete, and is everything cited appropri
 
 -->
 
-This paper announces OpenOA version 2.0.0, which was released in August of 2020. Originally released to the public in 2018, OpenOA is implemented in the Python programming language [cite python] and provides a framework and set of convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate big data analysis of operational data sets from wind energy plants. With over 50 stars on github, a dozen contributors, and an active Github issues forum, OpenOA is a mature and active project that provides a high level interface to practical problems in this field.
+This paper announces OpenOA version 2.0.0, which was released in August of 2020. Originally released to the public in 2018 [@osti_1478526], OpenOA is an open source framework for operational data analysis of wind energy plants, implemented in the Python programming language [@Rossum2009]. OpenOA provides a common data model, high level analysis workflows, and low-level convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate anlytics workflows on operational data sets. With over 50 stars on Github, a dozen contributors, and an active issues forum, OpenOA is becoming a mature project that provides a high level interface to solve practical problems in the wind energy industry.
 
 # Statement of Need
 
-Operational analyses are 
+OpenOA was created by and primarily developed by researchers at the National Renewable Energy Laboratory [^nrelwebsite] (NREL) through the Performance, Risk Uncertainty Finance [^wp3website] (PRUF) project, when the team recognized the need to compute a 20-year long term correced AEP metric from operational data as part of a benchmarking study [WP3 benchmark citation]. Due to restrictions on the input SCADA data, the team recognized that open source publication of the code was necessary to foster trust in the results by the participants of the benchmarking study. Furthermore, after talking with our industry partners, it became clear that there was no industry standard code for computing a 20-year long term corrected AEP.
 
-OpenOA was created by and primarily developed by researchers at the National Renewable Energy Laboratory [nrel footnote] through the Performance, Risk Uncertainty Finance (PRUF) project. 
+[^nrelwebsite]: \url{https://nrel.gov}
+[^wp3website]: \url{https://a2e.energy.gov/projects/wp3}
 
-OpenOA has also been used to calculate long-term operational AEP from over 470 wind farms in the US to assess correlation between uncertainty components [Bodini2020].
+**Figure 1: Overview of Operational Analysis**
 
-Reproducible research is an important goal of OpenOA.
+Operational analysis involves obtaining time-series data from an industrial plant's SCADA system, performing ETL and QC processes on these data, and then computing various metrics that might inform decisions by the plant operator. Since its inception, OpenOA has been used in several published studies at NREL. In one study, it is used to calculate long-term operational AEP from over 470 wind farms in the US to assess correlation between uncertainty components [@Bodini2020].
 
-OpenOA serves as a conduit to deliver state-of-the-art analysis methods between researchers and practitioners.
-
-We believe the reproducibility and efficiency of scientific research will increase if common, open source tools such as OpenOA are utilized.
-
-[Diagram of OA analysis]
+By forming an open source software project, OpenOA hopes to improve the reproducibility of research in this field, provide benchmark implementations of commonly performed data transformation and analysis tasks (to lower the barrier to entry), and finally to serve as a conduit that can deliver state-of-the-art analysis methods from researchers to practitioners.
 
 # Summary
 
-OpenOA V2 is released as a Python package and is freely available under a business-friendly, open-source, BSD license. OpenOA contains documentation, worked out examples in Jupyter notebooks, and an example dataset from Engie [cite engie].
[…]
