### 2020-08-18 #2171 @sjpfenninger
@jordanperr you are right, the version string is at v2.0; I was only looking at `CHANGELOG.md` which is _not_ updated for v2.0
