### 2020-08-18 bc781d1
advancing changelog to v2.0.0
