### 2020-08-18 #2171 @jordanperr
@sjpfenninger Oops, that was an oversight. Thanks for catching it. I have updated the changelog to reflect the release of v2.0.0.
