### 2020-08-19 NREL/OpenOA#92 @gschivley
Missing packages in installation
It looks like h5pyd, which is imported in operational_analysis/methods/quality_check_automation.py, isn't in setup.py or requirements.txt.

Since the examples are provided as jupyter notebooks and use matplotlib to show results, I'd also suggest that those packages should be either included in the default installation or called out as needed for the examples. matplotlib is included in the requirements.txt - which is referenced under [development](https://openoa.readthedocs.io/en/latest/install.html#development) installation - but jupyter isn't included anywhere. These are less critical than h5pyd though.

Part of review (openjournals/joss-reviews#2171)
