### 2020-08-21 NREL/OpenOA#93 @gschivley
Test instructions not working
I get an error when I try to run tests with the flag `-o python_files=test/*.py`.

```
❯ pytest -o python_files=test/*.py --cov=operational_analysis
zsh: no matches found: python_files=test/*.py
```

Is there something I'm missing? I can run tests without the flag or by doing `python setup.py test`.

Part of review (openjournals/joss-reviews#2171)
