### 2020-08-24 NREL/OpenOA#93 @jordanperr
Hi @gschivley, I have not run into this error yet. The python_files flag sets a relative path from the working directory to the test files, using a glob syntax to match multiple files.

If you navigate to ./test from the current directory, do you see these files? https://github.com/NREL/OpenOA/tree/master/test
