### 2020-08-25 NREL/OpenOA#78 @jordanperr
I'm closing this issue and we'll transition the discussion of pytest flags to #93
