### 2020-08-25 304fd89
updated dependency and docs for issue #92. Some uncommitted changes to the paper made it in here as well
