+++ paper/paper.bib
@@ -183,3 +183,19 @@ year = {2018},
 month = {2},
 note = 
 }
+
+
+@article{Craig2018,
+	doi = {10.1088/1742-6596/1037/5/052021},
+	url = {https://doi.org/10.1088%2F1742-6596%2F1037%2F5%2F052021},
+	year = 2018,
+	month = {jun},
+	publisher = {{IOP} Publishing},
+	volume = {1037},
+	pages = {052021},
+	author = {Anna Craig and Mike Optis and Michael Jason Fields and Patrick Moriarty},
+	title = {Uncertainty quantification in the analyses of operational wind power plant performance},
+	journal = {Journal of Physics: Conference Series},
+	abstract = {In the present work, we examine the variation introduced in the evaluation of an operating plant’s wind power production as a result of the choices analysts make in the processing of the operational data. For this study, an idealized power production for individual turbines over an operational period was predicted by fitting power curves to the turbine production data collected during expected operation (that is, without curtailment or availability losses). A set of 240 possible methods were developed for (a) defining what data represented expected operation and (b) modeling the power curve. The spread in the idealized power production as predicted by the different methods was on average almost 3% for the 100 turbines considered. Such significant variation places a lower bound on the precision with which analysts may employ such data as benchmarks for calibration of their energy estimation processes and limits the potential for identification of refinements to the energy estimation models for improved accuracy.}
+}
+
