+++ requirements.txt
@@ -7,6 +7,7 @@ pygam>=0.8.0
 scipy>=1.1.0
 statsmodels>=0.11
 tqdm>=4.28.1
+h5pyd>=0.7.0
 
 # Not included in docs requirements
 pytest>=5.4.2
