### 2020-08-25 NREL/OpenOA#92 @jordanperr
Thanks for catching these issues.

In branch joss-review:

- h5py has been added to setup.py and requirements.txt
- matplotlib is now included in requirements.txt and setup.py, since it is required to run core functionality of OpenOA.
- We consider jupyter like an IDE or text editor that a developer may choose to view the example notebooks. The notebooks can be opened using other tools (e.g., VSCode, Github's web interface, pandoc). I have added a line in the Readme suggesting Jupyter to open the example notebooks.
