### 2020-08-25 NREL/OpenOA#93 @RHammond2
Hi @jordanperr I was able to reproduce this error on my end. The zsh issue is more of a mask to the problem.

In the screenshot below I temporarily ignore the zsh error to see if I can find the underlying issue. This turns out to be even more confusing because it seems to suggest that no argument is being passed to the `-o/--override-ini` flag. I double checked thatf the testing files actually exist (they do) and that the argument we're passing is correct (they are) so it's a bit uncertain what this is stemming from.

Happy to keep digging in on this, but thought I should confirm this is able to be recreated.

Versions:
- python: 3.7.7
- pytest: 5.4.3
- pytest-cov: 2.10.0
