### 2020-08-25 f54b30c
update documentation and GH Actions workflows
