+++ readme.md
@@ -88,13 +88,13 @@ unzip examples/data/la_haute_borne.zip -d examples/data/la_haute_borne/
 Tests are written in the Python unittest framework and are runnable using pytest. To run all tests with code coverage reporting:
 
 ```
-pytest -o python_files=test/*.py --cov=operational_analysis
+pytest --cov=operational_analysis
 ```
 
 To run unit tests only:
 
 ```
-pytest -o python_files=test/test_*.py --cov=operational_analysis
+pytest --ignore=test/regression/ --cov=operational_analysis
 ```
 
 #### Documentation
