+++ readme.md
@@ -117,6 +117,7 @@ Nathan Agarwal,
 Nicola Bodini,
 Anna Craig,
 Jason Fields,
+Rob Hammond,
 Travis Kemper,
 Joseph Lee,
 Monte Lunacek,
