### 2020-08-26 NREL/OpenOA#93 @gschivley
@jordanperr Yes, I do have the test files and was able to run using

```
>>> python setup.py test
```

I think that #96 will take care of the issue by adding the `pytest.ini` file. Thanks @RHammond2.
