### 2020-08-31 NREL/OpenOA#93 @jordanperr
I have merged in PR96 to develop. This should resolve the issue.
