### 2020-09-15 4aab485
removing geopandas installation from setup.py and moving import statement in AssetData class
