### 2020-09-21 2929450
minor documentation updates
