### 2020-09-21 9102a0f
Unfreezing sphinx dependencies
