### 2020-09-25 #2171 @jordanperr
@sjpfenninger, We have completed the issues reported by the JOSS reviewers in our issue tracker and have merged these into develop (and at this point, they are also in Master). The most recent version of the paper is still in the joss-review branch, with an open pull request in to develop.

There are still some minor things to do on the paper itself. The major challenge is that we've had a new major version release of OpenOA between the time the paper was written and today.

How should we proceed? Has the repository satisfied the review requirements?
