### 2020-09-27 c721486
tweaks to paper. cleaning up citations
