+++ paper/paper_v1.md
@@ -1,187 +1,0 @@
----
-title: 'OpenOA: An Open-Source Codebase For Operational Analysis of Wind Farms'
-tags:
-  - Python
-  - wind energy
-  - operational analysis
-  - data analysis
-  - standardization
-authors:
-  - name: Jordan Perr-Sauer
-    orcid: 0000-0003-0872-7098
-    affiliation: 1 # (Multiple affiliations must be quoted)
-  - name: Mike Optis
-    orcid: 0000-0001-5617-6134
-    affiliation: 1
-  - name: Jason M. Fields
-    affiliation: 1
-  - name: Nicola Bodini
-    orcid: 0000-0002-2550-9853
-    affiliation: 1
-  - name: Lindy Williams
-    affiliation: 1
-  - name: Joseph C.Y. Lee
-    orcid: 0000-0003-1897-6290
-    affiliation: 1
-  - name: Austin Todd
-    affiliation: 1
-  - name: Caleb Phillips
-    affiliation: 1
-  - name: Monte Lunacek
-    affiliation: 1
-  - name: Anna Craig
-    affiliation: 1
-  - name: Travis Kemper
-    affiliation: 1
-  - name: Nathan Agarwal
-    orcid: 0000-0002-2734-5514
-    affiliation: 1
-  - name: Shawn Sheng
-    affiliation: 1
-  - name: Eric Simley
-    affiliation: 1
-  - name: John Meissner
-    affiliation: 1
-affiliations:
- - name: National Renewable Energy Laboratory, Golden, CO, USA
-   index: 1
-date: 20 December 2019
-bibliography: paper.bib
----
-
-# Summary
-
-OpenOA is an open source Python package which implements operational analysis (OA) methods for wind energy plants.
-The goal of OpenOA is collaboration and methods sharing in a wind industry that has historically been very protective of methods and data [@McCann2018].
-Over time, we hope that OpenOA will become a reference implementation for OA methods.
-
-The development of OpenOA started internally at the National Renewable Energy Laboratory (NREL) to support the lab's
-efforts in the Wind Plant Performance and Prediction (WP3) benchmarking initiative [^wp3website], which is a key risk reduction
-activity of the Performance, Risk, Uncertainty, and Finance project under the Atmosphere to Electrons initiative.
-OpenOA was originally scoped to calculate the operational annual energy production (AEP) of case study wind power plants, providing a baseline to which EYA bias can be measured.
-OpenOA has since expanded its scope to support additional types of analyses including turbine performance and electrical losses and to support data from wind plants outside the initial set of test projects.
-
-Released publicly in September 2018, the OpenOA repository contains numerous examples demonstrated in Jupyter notebooks along with publicly available data which can be used to run the built in unit and integration tests. The software has also been used in study on the impact of analyst choice in the outcome of OAs [@craig2018].
-
-[^wp3website]: \url{https://a2e.energy.gov/projects/wp3}
-
-## Importance of Operational Analysis
-Operational analysis uses collected data from wind farms to perform assessments ranging from the diagnosis of
-faults and underperformance, benchmarking of performance improvements (e.g., wind sector management,
-vortex generators), long-term estimates of AEP, and building
-statistical or physics-based models for various applications (e.g.,, wake model validation, wind power
-forecasting). Data sources include the wind farm revenue meter, turbine supervisory, control and data
-acquisition (SCADA) systems, on-site meteorological towers, and modeled atmospheric data (e.g., reanalysis data).
-
-# Built In Analysis Methods
-
-OpenOA includes three built in analysis methods, which operate on PlantData objects to produce analysis results. The three methods are: Long term corrected anualized energy production (AEP), Electrical Losses, and Turbine Ideal Energy. 
-
-## Long Term Corrected AEP
-
[…]
