+++ paper/paper.bib
@@ -1,9 +1,3 @@
-@misc{A2EWebsite,
-  title = {Atmosphere to Electrons Website},
-  howpublished = {\url{https://a2e.energy.gov}},
-  note = {Accessed: 2019-10-01}
-}
-
 @misc{WP3Website,
   title = {Wind Plant Performance Prediction (WP3) Benchmarking Project},
   howpublished = {\url{https://a2e.energy.gov/projects/wp3}},
@@ -19,20 +13,6 @@ title = {{Topical Expert Meeting {\#}92 on Wind Energy and Digitalization}},
 year = {2018}
 }
 
-
-@article{craig2018,
-	doi = {10.1088/1742-6596/1037/5/052021},
-	url = {https://doi.org/10.1088%2F1742-6596%2F1037%2F5%2F052021},
-	year = 2018,
-	month = {jun},
-	publisher = {{IOP} Publishing},
-	volume = {1037},
-	pages = {052021},
-	author = {Anna Craig and Mike Optis and Michael Jason Fields and Patrick Moriarty},
-	title = {Uncertainty quantification in the analyses of operational wind power plant performance},
-	journal = {Journal of Physics: Conference Series},
-	abstract = {In the present work, we examine the variation introduced in the evaluation of an operating plant’s wind power production as a result of the choices analysts make in the processing of the operational data. For this study, an idealized power production for individual turbines over an operational period was predicted by fitting power curves to the turbine production data collected during expected operation (that is, without curtailment or availability losses). A set of 240 possible methods were developed for (a) defining what data represented expected operation and (b) modeling the power curve. The spread in the idealized power production as predicted by the different methods was on average almost 3% for the 100 turbines considered. Such significant variation places a lower bound on the precision with which analysts may employ such data as benchmarks for calibration of their energy estimation processes and limits the potential for identification of refinements to the energy estimation models for improved accuracy.}
-}
 @article{lunacek2018,
   author={Monte Lunacek and M. Jason Fields and Anna Craig and Joseph C. Y. Lee and John Meissner and Caleb Philips and Shuangwen Sheng and Ryan King},
   title={Understanding Biases in Pre-Construction Estimates},
@@ -46,23 +26,6 @@ year = {2018}
   abstract={The pre-construction energy generation of a wind farm (P50) is difficult to estimate and evaluate. This paper presents a methodology to measure the accuracy of the p50 prediction, which we call the Historical Validation Survey (HVS), for several wind farms in the continental United States. Our results indicate that there is a bias between predicted and measured energy, even when controlling for factors like grid curtailment and resource variability. We also find that our results depend on the assumptions we make during analysis, which we quantify with a sensitivity analysis. This method allows the estimation of uncertainty we have in our findings. When we account for reasonable ranges of model assumptions, we find that, in the most optimistic case, there is still a bulk −5.5% bias when estimating pre-construction energy generation. When controlling for grid curtailment this number reduces to a range of −3.5 to −4.5%.}
 }
 
-@TECHREPORT{lindvall2016,
-  AUTHOR =        {Lindvall, J. and Hansson, J. and Undheim, O. and Vindteknikk, J.},
-  TITLE =         {Post-construction production assessment of wind farms},
-  NUMBER =        {2016:297},
-  INSTITUTION =   {Energyforsk},
-  YEAR  =         {2016},
-  ISBN =          {978-91-7673-297-7},
-}
-
-@mastersthesis{khatab2017,
-  author       = {A Mouez Khatab},
-  title        = {Performance Analysis of Operating Wind Farms},
-  school       = {Uppsala University},
-  year         = 2017,
[…]

[…]
