### 2020-09-27 #2171 @jordanperr
@sjpfenninger The paper is now ready aside from some very minor citation formatting issues which I will continue to work on. Please proceed with the review. The whole team is excited and eager to get this across the finish line! Thanks for your perseverance!
