### 2020-09-28 6220c64
update docstrings
