### 2020-09-29 52739d3
fix typo and update imputation methods for arrays
