+++ operational_analysis/toolkits/imputing.py
@@ -7,42 +7,44 @@ import pandas as pd
 from tqdm import tqdm
 
 
-def correlation_matrix_by_id_column(df, align_col, id_col, value_col):
+def correlation_matrix_by_id_column(align_col, id_col, value_col):
     """Create a correlation matrix between different assets in a data frame
 
     Args:
-        df(:obj:`pandas.DataFrame`): input data frame
-        align_col(:obj:`str`): name of column in <df> on which different assets are to be aligned
-        id_col(:obj:`str`): the column distinguishing the different assets
-        value_col(:obj:`str`): the column containing the data values to be used when assessing correlation
+        align_col(:obj:`array-like`): array of different assets are to be aligned
+        id_col(:obj:`array-like`): array distinguishing the different assets
+        value_col(:obj:`array-like`): array containing the data values to be used when assessing correlation
 
     Returns:
         :obj:`pandas.DataFrame`: Correlation matrix with <id_col> as index and column names
     """
-    assets = df[id_col].unique()  # List different assets in <id_col>
+    assets = np.unique(id_col)  # List different assets in <id_col>
     corr_df = pd.DataFrame(index=assets, columns=assets, dtype=float,
                            data=np.nan)  # Define correlation matrix that is set to NaN by default
+    df = pd.DataFrame(np.column_stack((id_col, align_col, value_col)), columns=["id_col", "align_col", "value_col"])
+    df.value_col = df.value_col.astype(float)
 
     for t in assets:  # Loop through assets
         assets = assets[assets != t]  # Remove the reference asset so we don't loop over it again
         for s in assets:  # Loop through remaining assets
-            x_df = df.loc[df[id_col] == t, [align_col, value_col]]  # Asset 't'
-            y_df = df.loc[df[id_col] == s, [align_col, value_col]]  # Asset 's'
+            ix_t = id_col == t
+            x_df = df.loc[df.id_col == t, ["align_col", "value_col"]]  # Asset 't'
+            y_df = df.loc[df.id_col == s, ["align_col", "value_col"]]  # Asset 's'
 
-            merged_df = x_df.merge(y_df, on=align_col).dropna()  # Merge the two on <align_col>, drop any rows with NaN
+            merged_df = x_df.merge(y_df, on="align_col").dropna()  # Merge the two on <align_col>, drop any rows with NaN
 
             # If merged data frame is empty or has only 1 entry, assign NaN correlation
             if (merged_df.empty) | (merged_df.shape[0] < 2):
                 corr_df.loc[t, s] = np.nan
             else:  # Now assign correlations
-                corr_df.loc[t, s] = np.corrcoef(merged_df[value_col + '_x'], merged_df[value_col + '_y'])[0, 1]
+                corr_df.loc[t, s] = np.corrcoef(merged_df['value_col_x'], merged_df['value_col_y'])[0, 1]
 
             corr_df.loc[s, t] = corr_df.loc[t, s]  # entry (t,s) is the same as (s,t)
 
     return corr_df
 
 
-def impute_data(target_data, target_value_col, ref_data, ref_value_col, align_col,
+def impute_data(target_value_col, target_align_col, ref_value_col, ref_align_col, index_col=None,
                 method='linear'):  # ADD LINEAR FUNCTIONALITY AS DEFAULT, expection otherwise
     """Replaces NaN data in a target Pandas series with imputed data from a reference Panda series based on a linear
     regression relationship.
@@ -55,26 +57,31 @@ def impute_data(target_data, target_value_col, ref_data, ref_value_col, align_co
     4. Return the imputed results as well as the index matching the target data frame
 
     Args:
-        target_data(:obj:`pandas.DataFrame`): the data frame containing NaN data to be imputed
-        target_value_col(:obj:`str`): the name of the column in <target_data> to be imputed
[…]
