+++ test/test_imputing_toolkit.py
@@ -1,4 +1,5 @@
 import unittest
+import pytest
 
 import numpy as np
 import pandas as pd
@@ -91,36 +92,56 @@ class SimpleFilters(unittest.TestCase):
 
     def test_correlation_matrix_by_id_column(self):
         # Test 1, make sure a simple correlation of two assets works
-        y = imputing.correlation_matrix_by_id_column(self.test_df, 'time', 'id', 'data')
+        y = imputing.correlation_matrix_by_id_column(self.test_df.time, self.test_df.id, self.test_df.data)
         nptest.assert_array_almost_equal(y, np.array([[np.nan, 0.970166], [0.970166, np.nan]]), decimal=4)
 
         # Test 2, if no overlapping data are present, make sure correlation matrix is all NaN
-        y2 = imputing.correlation_matrix_by_id_column(self.test9_df, 'time', 'id', 'data')
+        y2 = imputing.correlation_matrix_by_id_column(self.test9_df.time, self.test9_df.id, self.test9_df.data)
         nptest.assert_array_equal(y2, np.array([[np.nan, np.nan], [np.nan, np.nan]]))
 
     def test_impute_data(self):
         # Test 1, make sure single NaN is imputed properly and that the index is correct
-        y = imputing.impute_data(self.test_df.loc[self.test_df.id == 'a'], 'data',
-                                 self.test_df.loc[self.test_df.id == 'b'], 'data', 'time')
+        target = self.test_df.loc[self.test_df.id == 'a']
+        ref = self.test_df.loc[self.test_df.id == 'b']
+        y = imputing.impute_data(target.data, target.time, ref.data, ref.time, target.index.values)
         nptest.assert_almost_equal(np.float(y.values[1]), np.float(2.989779), decimal=4)
 
         # Test 2, make sure only the first NaN entry is imputed
-        y2 = imputing.impute_data(self.test2_df.loc[self.test2_df.id == 'a'], 'data',
-                                  self.test2_df.loc[self.test2_df.id == 'b'], 'data', 'time')
+        target = self.test2_df.loc[self.test2_df.id == 'a']
+        ref = self.test2_df.loc[self.test2_df.id == 'b']
+        y2 = imputing.impute_data(target.data, target.time, ref.data, ref.time, target.index.values)
         nptest.assert_almost_equal(np.float(y2.loc['c']), np.float(3.874429), decimal=4)
         nptest.assert_equal(y2.loc['b'], np.nan)
 
         # Test 3, make sure no data is imputed when no NaN are present
-        y3 = imputing.impute_data(self.test3_df, 'data1', self.test4_df, 'data2', 'align')
+        y3 = imputing.impute_data(
+            self.test3_df.data1,
+            self.test3_df["align"],
+            self.test4_df.data2,
+            self.test4_df["align"],
+            self.test3_df.index.values
+        )
         nptest.assert_array_almost_equal(y3.to_numpy(), self.test3_df['data1'].to_numpy())
 
         # Test 4, make sure two NaNs are imputed when there 3 total but reference data only matches 2
-        y4 = imputing.impute_data(self.test5_df, 'data1', self.test4_df, 'data2', 'align')
+        y4 = imputing.impute_data(
+            self.test5_df.data1,
+            self.test5_df["align"],
+            self.test4_df.data2,
+            self.test4_df["align"],
+            self.test5_df.index
+        )
         nptest.assert_array_almost_equal(y4.loc[['c', 'd']], np.array([2.0, 4.0]), decimal=4)
 
         # Test 5, make sure exception is thrown if no valid data are available
         def func():  # Function to return exception
-            imputing.imputing.impute_data(self.test5_df, 'data1', self.test7_df, 'data2', 'align')
[…]
