### 2020-09-29 5750dc2
Update plant_analysis.py
