+++ operational_analysis/methods/plant_analysis.py
@@ -35,10 +35,8 @@ class MonteCarloAEP(object):
     analysis methods for estimating plant level operational AEP and uncertainty.
     
     The preprocessing should run in this order:
-
         1. Process revenue meter energy - creates monthly/daily data frame, gets revenue meter on monthly/daily basis, and adds
            data flag
-
         2. Process loss estimates - add monthly/daily curtailment and availabilty losses to monthly/daily data frame
         
         3. Process reanalysis data - add monthly/daily density-corrected wind speeds, temperature (if used) and wind direction (if used)
@@ -47,12 +45,11 @@ class MonteCarloAEP(object):
         4. Set up Monte Carlo - create the necessary Monte Carlo inputs to the OA process
         
         5. Run AEP Monte Carlo - run the OA process iteratively to get distribution of AEP results
-
     The end result is a distribution of AEP results which we use to assess expected AEP and associated uncertainty
     """
 
     @logged_method_call
-    def __init__(self, plant, reanal_products=["merra2", "ncep2", "erai"], uncertainty_meter=0.005, uncertainty_losses=0.05,
+    def __init__(self, plant, reanal_products=["merra2", "ncep2", "erai","era5"], uncertainty_meter=0.005, uncertainty_losses=0.05,
                  uncertainty_windiness=(10, 20), uncertainty_loss_max=(10, 20),
                  uncertainty_nan_energy=0.01, time_resolution = 'M', reg_model = 'lin', ml_setup_kwargs={},
                  reg_temperature = False, reg_winddirection = False):
@@ -408,14 +405,12 @@ class MonteCarloAEP(object):
     def process_revenue_meter_energy(self):
         """
         Initial creation of monthly data frame:
-
             1. Populate monthly/daily data frame with energy data summed from 10-min QC'd data
             2. For each monthly/daily value, find percentage of NaN data used in creating it and flag if percentage is
                greater than 0
             
         Args:
             (None)
-
         Returns:
             (None)
         """
@@ -499,7 +494,6 @@ class MonteCarloAEP(object):
     def process_reanalysis_data(self):
         """
         Process reanalysis data for use in PRUF plant analysis:
-
             - calculate density-corrected wind speed and wind components
             - get monthly/daily average wind speeds and components
             - calculate monthly/daily average wind direction
@@ -642,19 +636,19 @@ class MonteCarloAEP(object):
 
         # If valid data hasn't yet been stored in dictionary, determine the valid data
         df = self._aggregate.df
-                
+        
         # First set of filters checking combined losses and if the Nan data flag was on
         df_sub = df.loc[
             ((df['availability_pct'] + df['curtailment_pct']) < self._run.loss_threshold) & (df['nan_flag'] == False),:]
         
         # Set maximum range for using bin-filter, convert from MW to GWh
-        plant_capac = self._plant._plant_capacity/1000 * self._hours_in_res
+        plant_capac = self._plant._plant_capacity/1000. * self._hours_in_res
 
-        # Apply range filter to wind speed
+        #Apply range filter to wind speed
         df_sub = df_sub.assign(flag_range=filters.range_flag(df_sub[reanal], below = 0, above = 40))
-        # Apply frozen/unresponsive sensor filter
+        #Apply frozen/unresponsive sensor filter
         df_sub.loc[:,'flag_frozen'] = filters.unresponsive_flag(df_sub[reanal], threshold = 3)
-        # Apply window range filter
+        #Apply window range filter
         df_sub.loc[:,'flag_window'] = filters.window_range_flag(window_col = df_sub[reanal], 
                                                                     window_start = 5., 
[…]
