+++ test/int_pruf_plant_analysis.py
@@ -227,7 +227,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
  
     def check_simulation_results_lin_monthly(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.41, 8.29, 1.30, 3.57, 0.09, 3.57]
+        expected_results = [12.41, 2.29, 1.30, 3.57, 0.09, 3.57]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
