+++ setup.py
@@ -53,7 +53,8 @@ setup(name='OpenOA',
                         "statsmodels",
                         "scikit_learn",
                         "EIA-python",
-                        "requests"],
+                        "requests",
+                        "pyproj"],
       tests_require=['pytest', 'pytest-cov'],
       python_requires='>=3.6'
       )
