### 2020-10-08 11f632a
add shapely as dependency
