### 2020-10-12 7dacd58
Update python-ci.yml
Includes Windows to ensure dependencies work on all systems.
