+++ operational_analysis/methods/plant_analysis.py
@@ -83,6 +83,7 @@ class MonteCarloAEP(object):
         self.outlier_filtering = {}  # Combinations of outlier filter results
         self.long_term_sampling = {}  # Combinations of long-term reanalysis data sampling
         self.long_term_sampling_sd = {}  # Combinations of long-term reanalysis data sampling stdev of wind speed
+        self.opt_model = {} # Optimized ML model hyperparameters for each reanalysis product
         
         # Define relevant uncertainties, data ranges and max thresholds to be applied in Monte Carlo sampling
         self.uncertainty_meter = np.float64(uncertainty_meter)
@@ -776,12 +777,24 @@ class MonteCarloAEP(object):
         # Machine learning models
         else: 
             ml = MachineLearningSetup(self.reg_model, **self.ml_setup_kwargs)
-            ml.hyper_optimize(np.array(reg_data[:,0:-1]), reg_data[:,-1], n_iter_search = 5, report = False, cv = KFold(n_splits = 2))
-            predicted_y = ml.random_search.predict(np.array(reg_data[:,0:-1]))
+            # Memoized approach for optimized hyperparameters
+            if self._run.reanalysis_product in self.opt_model:
+                self.opt_model[(self._run.reanalysis_product)].fit(np.array(reg_data[:,0:-1]), reg_data[:,-1])
+                predicted_y = self.opt_model[(self._run.reanalysis_product)].predict(np.array(reg_data[:,0:-1]))
+                
+                self._r2_score[n] = r2_score(reg_data[:,-1], predicted_y)
+                self._mse_score[n] = mean_squared_error(reg_data[:,-1], predicted_y)
+                return self.opt_model[(self._run.reanalysis_product)]
+        
+            else: # optimize hyperparameters once for each reanalysis product  
+                ml.hyper_optimize(np.array(reg_data[:,0:-1]), reg_data[:,-1], n_iter_search = 20, report = False, cv = KFold(n_splits = 5))
+                # Store optimized hyperparameters for each reanalysis product
+                self.opt_model[(self._run.reanalysis_product)] = ml.opt_model
+                predicted_y = ml.random_search.predict(np.array(reg_data[:,0:-1]))
             
-            self._r2_score[n] = r2_score(reg_data[:,-1], predicted_y)
-            self._mse_score[n] = mean_squared_error(reg_data[:,-1], predicted_y)
-            return ml.random_search
+                self._r2_score[n] = r2_score(reg_data[:,-1], predicted_y)
+                self._mse_score[n] = mean_squared_error(reg_data[:,-1], predicted_y)
+                return ml.random_search
 
     @logged_method_call
     def run_AEP_monte_carlo(self):
