### 2020-10-12 b1fd9fc
Update python-ci.yml
Forces bash shell usage to be compatible across OS options.
