+++ .github/workflows/python-ci.yml
@@ -21,16 +21,18 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
-      shell: bash
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
+      shell: bash
     - name: Unzip example data
       run: |
         yes | unzip examples/data/la_haute_borne.zip -d examples/data/la_haute_borne/
+      shell: bash
     - name: Test with pytest
       run: |
-        pytest --cov=operational_analysis --cov-report=xml
+        pytest --cov=operational_analysis --cov-report=xml'
+      shell: bash
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
