+++ .github/workflows/python-ci.yml
@@ -25,7 +25,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        yes
       shell: bash
     - name: Unzip example data
       run: |
