+++ .github/workflows/python-ci.yml
@@ -32,7 +32,7 @@ jobs:
       shell: bash
     - name: Test with pytest
       run: |
-        pytest --cov=operational_analysis --cov-report=xml'
+        pytest --cov=operational_analysis --cov-report=xml
       shell: bash
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
