### 2020-10-13 b49ac2d
replace geopandas with pyproj (#107)
* replace `GeoPandas` with `pyproj` and `Shapely`

* adding `pyproj` and `shapely` to `setup.py` and `requirements.txt`

* Update `python-ci.yml`to include Windows tests

* update software version to 2.0.1

Co-authored-by: Rob <robert.hammond@nrel.gov>
