### 2020-10-13 da04091
Update AssetData docstring
Removed any lingering references to `GeoPandas`.
