### 2020-10-13 fd542fc
Remove GDAL/Windows comments
