+++ readme.md
@@ -104,6 +104,7 @@ Nathan Agarwal,
 Nicola Bodini,
 Anna Craig,
 Jason Fields,
+Rob Hammond,
 Travis Kemper,
 Joseph Lee,
 Monte Lunacek,
