+++ test/int_pruf_plant_analysis.py
@@ -293,6 +293,5 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
     def tearDown(self):
         pass
 
-
 if __name__ == '__main__':
     unittest.main()
