### 2020-10-23 6605917
add binder links to main page and fix typo
