+++ sphinx/index.rst
@@ -29,6 +29,9 @@ Analysis routines are grouped by purpose into toolkits - which provide an abstra
 computations, and methods - which provide higher level wind industry specific API. In addition to these provided modules,
 anyone can write their own, which is intended to provide natural growth of tools within this framework.
 
+To interact with how each of these components of OpenOA are used, please visit our examples notebooks on
+`Binder <https://mybinder.org/v2/gh/NREL/OpenOA/master?filepath=examples>`_, or view them statically on the
+`examples page <examplesout>`_.
 
 .. toctree::
     :maxdepth: 2
