### 2020-11-05 366da74
fixing pyproj version to 2.6.1
pyproj 3.0.0 breaks with windows pip installation.
