+++ .github/workflows/python-ci.yml
@@ -23,7 +23,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip
+        python -m pip install --upgrade pip==20.2.3
         pip install -r requirements.txt
       shell: bash
     - name: Unzip example data
