### 2020-11-05 2f18e3b
Add only-binary flag to pyproj
