### 2020-11-06 be22c77
Update plant_analysis.py
