### 2020-11-10 aaea13e
updating for pyproj oddity on rtd build
