### 2020-11-20 da52164
Update plant_analysis.py
