### 2020-11-20 82fc68b
Update plant_analysis.py
