### 2020-12-01 #2171 @sjpfenninger
@ckaldemeyer @SabineHaas Would you be interested to jump in and help review this submission to JOSS?
