### 2020-12-11 NREL/OpenOA#92 @gschivley
@jordanperr it looks like h5py was added to requirements.txt but not setup.py in the joss-review branch. And will these changes be merged into the main branch?
