### 2020-12-11 NREL/OpenOA#113 @gschivley
Small example bug
In the 00_toolkit_examples notebook, use of 'True' instead of `True` for a boolean mask (cell 7) causes an error.

![image](https://user-images.githubusercontent.com/10373332/101917858-aee04a80-3b96-11eb-9306-6af4a4daa226.png)

Part of review (openjournals/joss-reviews#2171)
