### 2020-12-11 NREL/OpenOA#114 @gschivley
Install fails with Python 3.9
The installation instructions say to do

```
conda create --name openoa-env python=3
```

which now installs python 3.9. I tried to install one final time (previous laptop died) to finish my review, and it looks like there are issues (maybe some missing wheels, maybe just on Mac?). I had to specify python 3.8 to install.

Part of review (openjournals/joss-reviews#2171)
