### 2020-12-14 NREL/OpenOA#114 @brynpickering
I also came across this issue on OSX, similarly solved by forcing `python=3.8`.

Note: this comment comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2171)
