### 2020-12-14 NREL/OpenOA#115 @brynpickering
Example notebooks package requirements
It would be nice to jump into the example notebooks without having to deal with `ModuleNotFound` errors once working in them. Would it be possible to have a requirements file in the examples directory, containing all the additional packages required in the example notebooks? It could also include jupyter itself (or, at least, ipykernel).

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2171)
