### 2020-12-14 NREL/OpenOA#116 @brynpickering
Missing h5pyd package requirement
The required dependency `h5pyd` is missing following a fresh install of OpenOA.

This causes an error in the file `operational_analysis/methods/quality_check_automation.py`. This error is not caught by tests since the file coverage is 0%. It may be a good idea to just have a test to try importing every function in the package, to avoid this issue in future.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2171)
