### 2020-12-14 NREL/OpenOA#115 @brynpickering
I see that the main missing requirement from the standard setup is matplotlib. matplotlib is in the top-level `requirements.txt`, but this then would install development-based packages that users looking at the examples probably don't care about.

Perhaps it isn't worth it for just jupyter and matplotlib in the examples directory requirements; happy for you to close this if you also think it isn't worth it.
