### 2020-12-15 NREL/OpenOA#114 @jordanperr
This seems to be an issue with many packages right now. Please proceed reviewing OpenOA assuming the Python 3.8 dependency - and we will work to figure out the long term solution for Python 3.9.
