### 2020-12-15 NREL/OpenOA#115 @jordanperr
I like the idea of a separate requirements file for the example notebooks. We can add Jupyter there, as well.
