### 2020-12-15 NREL/OpenOA#116 @jordanperr
Good catch. And no, we do not have any test coverage for this file, which is the root of the problem. We have had some internal discussions over whether or not to require h5pyd in the general requirements, since many users may not use the quality_check_automation.py module. However, I agree that h5pyd should be added to the new example requirements, since it is required to run the example notebooks.
