### 2020-12-15 16157e7
include examples requirements and instructions
