+++ examples/requirements.txt
@@ -1,0 +1,3 @@
+h5pyd
+jupyterlab
+matplotlib
