### 2020-12-15 NREL/OpenOA#116 @brynpickering
Hmm, since this is a requirement of the package (i.e. it is explicitly imported in one of the package modules, even if that module is likely to be seldom used), it should probably be an installation requirement.
