### 2020-12-15 NREL/OpenOA#118 @RHammond2
pre-commit hooks to enable easier and more consistent contributions
Per our meeting I'd like to propose adding in [`pre-commit`](https://pre-commit.com/) hooks into the repository. This would help ensure that as new contributors come on board or as established contributors keep contributing we maintain at least a loose set of standards for how code should look. Below are some recommendations for what we could include or areas for further research.

1) [`black`](https://github.com/psf/black) for code formatting.
2) [`isort`](https://pycqa.github.io/isort/) for import sorting imports (works well with `black` when you align the settings).
3) So many settings for things ranging from json and yaml formatting to code linting: https://pre-commit.com/hooks.html.
   - pretty-format-json
   - check-json
   - check-yaml
   - end-of-file-fixer
   - trailing-whitespace
4) See the [`floris` repository](https://github.com/NREL/floris/blob/main/.pre-commit-config.yaml) for what we ended up doing there
