### 2020-12-15 NREL/OpenOA#119 @brynpickering
JOSS paper cleanup, content, and structure
The JOSS paper looks to cover everything I'd expect, but would benefit from a few cleanups:

- [x] Define abbreviations at least once (e.g. AEP, SCADA), or just avoid using them altogether. In the case of AEP, it is not a well known abbreviation (unlike e.g. PV for solar photovoltaic panels), so I'd have preference for it not being used at all.

- [x] Links to (undated) websites would look better as just hyperlinked text. I.e. don't bother including it in the list of references.

- [x] Reference to literature published by NREL should ideally use the 'shortauthor' format, so that the in-text citation reads `NREL` rather than `Laboratory`. Not sure how to make this happen in Markdown, but would be worth investigating.

- [x] Link to ReadtheDocs hasn't rendered. Instead, the generated PDF just shows `[^rtdwebsite]`

- [x] Some misplaced periods in `Version 2 of OpenOA implements three high level analysis methods. (1) Long term corrected AEP. (2) Electrical losses, and (3) Turbine level losses.`. Possibly should be `Version 2 of OpenOA implements three high level analysis methods: (1) Long term corrected AEP, (2) Electrical losses, and (3) Turbine level losses.`

### structure
- [x] The initial paragraph would be better placed in the summary (and the summary placed before the statement of need). If you want to retain the first paragraph, then you should place more importance on describing its high-level functionality than on the credentials of the github repository.

- [x] The paper seems to be missing any reference to other similar and commonly-used packages in the research area. This is a listed criterion in the [JOSS submission guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper)
- [x] You produce a number of informative figures in your example notebooks. Placing one or two of these in your paper might further aid readers in understanding the purpose of the software.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2171)

Edit: added in structure and content comments into this issue so everything is kept in one place.
