### 2020-12-17 #2171 @brynpickering
The submission looks to be in very good shape to me. I have opened a few issues on the OpenOA repo, but they're all relatively minor. Some comments from my end are below. Once the open issues are addressed, I'll complete my checklist.

## Installation

I opened two issues on this (https://github.com/NREL/OpenOA/issues/115, https://github.com/NREL/OpenOA/issues/116), since the current setup caused some confusion for me as to what packages are actually required by the package. These are now being jointly dealt with in an open PR (https://github.com/NREL/OpenOA/pull/117), and I think it's going in the right direction to having something understandable. A few minor comments that should not get in the way of this submission:

1. There is no reason to be quite so fearful of requiring 'too many' dependencies in the core installation, especially when those are such established packages like matplotlib. Keep in mind that those unfamiliar with package management will become very confused by 'ModuleNotFound' errors if confronted with them when trying to access a module in your package.
2. For greater ease, you may consider packaging this software for distribution on conda-forge. This would require moving the ENGIE data out of the package. To be honest, I would do this anyway: data and code are often better kept apart. Since the data is publicly available and accessible by API, you'd do better to access it on the fly.

## Code
All looks generally fine (I've not gone through it all line-by-line, of course). It could do with linting and applying [Black](https://github.com/psf/black), since there are some inconsistencies in the formatting. I see this is already on your radar (https://github.com/NREL/OpenOA/issues/118).

I would also caution against too much hardcoding within the code. In some instances, you could make better use of dependency functionality (e.g. `operational_analysis/toolkits/unit_conversion.py:convert_power_to_energy` could just resample the dataframe rather than expecting one of five hardcoded input time frequencies). In other cases, 'translations' could be kept inside e.g. YAML files, so that they're all in one place, configurable and out of the way of the Python code (e.g. scada_map in examples/project_ENGIE.py).

Finally, the `.gitignore` should include some more items: it's worth checking whether there are ones you don't notice because they're in your global gitignore (e.g. `.png` files).

## Paper
I have opened a 'meta' issue on this (https://github.com/NREL/OpenOA/issues/119), since there is some minor cleanup required, as well as tweaks to the content and structure I'd like to see. I would caution against mentioning stats related to the GitHub repository in the paper, since those can quickly become stale on the timescale the paper will be up and relevant. I'd also give more consideration to non-technical users of the software in the content since it currently feels more geared towards other software developers.
