### 2020-12-17 95d7acc
Undo changes in requirements.txt
Not sure why these changes were included in the first place..
