+++ examples/requirements.txt
@@ -1,3 +1,2 @@
-h5pyd
 jupyterlab
 matplotlib
