### 2020-12-17 1265134
Enhancement/add binder (#110)
* adding binder links to readme, examples, and documentation pages

* update so it runs in binder successfully

Co-authored-by: Rob <robert.hammond@nrel.gov>
