### 2020-12-17 9afc0ec
Bug Fix: missing requirements (#117)
* include examples requirements and instructions to install

* include h5pyd requirements at the top level

Co-authored-by: Rob <robert.hammond@nrel.gov>
