### 2020-12-17 NREL/OpenOA#121 @RHammond2
Dependency Management Improvements
It's come to our realization from our JOSS paper review that our dependency management system using `requirements.txt` and `setup.py` is getting a bit cumbersome.  It would be good to come up with a better way to handle the documentation site build requirements, developer requirements, and examples requirements than our traditional methods.

This could also be a good time to rethink how we handle packages required for specific, lesser used purposes, such as `h5pyd`. This is a lower priority item, but could be a good way into the package/workflows for anyone not directly involved in the maintenance of this project.
