### 2020-12-24 02d7a08
Fixing gross_por calculation
