### 2020-12-24 5b3daf0
Update plant_analysis.py
