### 2020-12-24 41c6052
Update plant_analysis.py
