### 2021-01-12 #2171 @SabineHaas
> @ckaldemeyer @SabineHaas Would you be interested to jump in and help review this submission to JOSS?

@sjpfenninger Unfortunately I couldn't find the time during these weeks, but my colleague @Bachibouzouk would be interested, if you still require an additional reviewer.
