### 2021-01-19 NREL/OpenOA#115 @jordanperr
This has been fixed in #117
