### 2021-01-19 NREL/OpenOA#116 @jordanperr
This was fixed in #117
