### 2021-01-19 NREL/OpenOA#113 @jordanperr
This was fixed in #110. Thanks, rob!
