### 2021-01-19 f240b39
fixing most of the small things from brynpickerings review. Added a figure, but have not described it yet.
