+++ paper/paper.md
@@ -75,27 +75,29 @@ References: Is the list of references complete, and is everything cited appropri
 
 -->
 
-OpenOA is an open source framework for operational data analysis of wind energy plants, implemented in the Python programming language [@Rossum2009]. Originally released to the public in 2018 [@osti_1478526], OpenOA is now actively developed through a public Github repository. OpenOA provides a common data model, high level analysis workflows, and low-level convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate anlytics workflows on operational data sets. With over 50 stars on Github, a dozen contributors, and an active issues forum, OpenOA is becoming a mature project that provides a high level interface to solve practical problems in the wind energy industry.
-
-# Statement of Need
-
-OpenOA was created by and primarily developed by researchers at the National Renewable Energy Laboratory (NREL) through the Performance, Risk Uncertainty Finance (PRUF) project, when the team recognized the need to compute a 20-year long term correced AEP metric from operational data as part of a benchmarking study [@WP3Website] [@lunacek2018]. Due to access restrictions on the input data, the team recognized that open source publication of the code was necessary to foster trust in the results by the participants of the benchmarking study. Furthermore, after talking with our industry partners, it became clear that there was no industry standard code for computing a 20-year long term corrected AEP.
-
-Operational analysis involves obtaining time-series data from an industrial plant's SCADA system, performing quality control processes on these data, and computing metrics that provide insight into the performance charactertistics of a wind plant. Since its inception, OpenOA has been used in several published studies at NREL. An early version of the code was used in [@Craig2018] to quantify the uncertainty in analyst choices. In one study, it is used to calculate long-term operational AEP from over 470 wind farms in the US to assess correlation between uncertainty components [@Bodini2020]. OpenOA is used in an upcoming paper analysing the gap between preconstruction estimates of energy production, called the P50 bias [@nrel2019].
-
-By forming an open source software project, OpenOA hopes to facilitate reproducibility of research in this field, provide benchmark implementations of commonly performed data transformation and analysis tasks (to lower the barrier to entry), and finally to serve as a conduit that can deliver state-of-the-art analysis methods from researchers to practitioners.
-
 # Summary
 
-OpenOA V2 is released as a Python package and is freely available under a business-friendly, open-source, BSD license. OpenOA contains documentation, worked out examples in Jupyter notebooks, and a corresponding example dataset from Engie Renewable's La Haute Borne Dataset [@EngieDataset].
+OpenOA is an open source framework for operational data analysis of wind energy plants, implemented in the Python programming language [@Rossum2009]. OpenOA provides a common data model, high level analysis workflows, and low-level convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate anlytics workflows on operational data sets. OpenOA contains documentation, worked out examples in Jupyter notebooks, and a corresponding example dataset from [Engie Renewable's La Haute Borne Dataset](https://opendata-renewables.engie.com/explore/dataset/d543716b-368d-4c53-8fb1-55addbe8d3ad/information).
 
-The typical user interfaces with OpenOA through its analysis methods. These are Python classes which conform to a common interface (e.g., they implement `__init__`, `prepare`, and `run` methods). Version 2 of OpenOA implements three high level analysis methods. (1) Long term corrected AEP. (2) Electrical losses, and (3) Turbine level losses. Uncertainty quantification is achieved in each analysis using a monte carlo approach. A more detailed description of these analyses are provided in the documentation [^rtdwebsite].
[…]
