+++ paper/paper.bib
@@ -26,10 +26,11 @@ year = {2018}
   abstract={The pre-construction energy generation of a wind farm (P50) is difficult to estimate and evaluate. This paper presents a methodology to measure the accuracy of the p50 prediction, which we call the Historical Validation Survey (HVS), for several wind farms in the continental United States. Our results indicate that there is a bias between predicted and measured energy, even when controlling for factors like grid curtailment and resource variability. We also find that our results depend on the assumptions we make during analysis, which we quantify with a sensitivity analysis. This method allows the estimation of uncertainty we have in our findings. When we account for reasonable ranges of model assumptions, we find that, in the most optimistic case, there is still a bulk −5.5% bias when estimating pre-construction energy generation. When controlling for grid curtailment this number reduces to a range of −3.5 to −4.5%.}
 }
 
-@techreport{nrel2019,
+@techreport{NRELPhase1Report,
   author = {National Renewable Energy Laboratory},
-  year = {in draft},
-  title = {Wind plant performance prediction (WP3) benchmark report},
+  shortauthor = {NREL},
+  year = {2020},
+  title = {Wind Plant Performance Prediction (WP3) Phase 1 Benchmark Report (In Draft)},
   institution = {Department of Energy}
 }
 
@@ -101,13 +102,6 @@ year = {2018}
   pages={90--95}, 
   year={2007}, 
   publisher={IEEE} 
-}
-
-
-@misc{EngieDataset,
-  title = {La Haute Borne Data (2013-2016)},
-  howpublished = {Engie Renewables. \url{https://opendata-renewables.engie.com/explore/dataset/d543716b-368d-4c53-8fb1-55addbe8d3ad/information}},
-  note = {Updated 9 October 2019}
 }
 
 
