### 2021-01-22 NREL/OpenOA#118 @jordanperr
I think this is a great idea. We used to enforce linting standards but found it was too cumbersome for novice contributors. There is still a section in contributing.md on this, recommending `pylint --max-line-length=120 operational_analysis` be run before committing.
