### 2021-01-22 NREL/OpenOA#114 @jordanperr
The main reason for install failure is that Shapely does not have binary wheels for Mac + Python 3.9. This is an issue specific to Macs running Python 3.9. There is an open ticket for this in the Shapely issue tracker. People in this thread suggest using brew to first install geos (which is the critical missing dependency for Shapely) before trying to install Shapely. This workaround will likely work for OpenOA, but I am have not tested it.

For now:
- Readme will be updated to recommend Python 3.8.
- I will leave this issue open until OpenOA supports Python 3.9 on all platforms.

Thanks again for reporting this.

https://github.com/Toblerity/Shapely/issues/1040
