### 2021-01-22 9ca3c21
capping supported python version at 3.8 until we can resolve the Shapely dependency issue.
