+++ paper/paper.bib
@@ -29,7 +29,7 @@ year = {2018}
 @techreport{NRELPhase1Report,
   author = {National Renewable Energy Laboratory},
   shortauthor = {NREL},
-  year = {2020},
+  year = {In progress},
   title = {Wind Plant Performance Prediction (WP3) Phase 1 Benchmark Report (In Draft)},
   institution = {Department of Energy}
 }
