### 2021-01-22 NREL/OpenOA#119 @jordanperr
Thank you for the suggestions, @brynpickering. The latest version in the joss-review branch addresses many of the changes. Specifically:

- I have integrated all of the smaller style and grammar changes that were suggested.
- Re-arranged the summary section to incorporate more language about uses of the software earlier on. Combined the abstract with the summary section as suggested.
- Ensure all abbreviations are defined
- Added a figure with example graphs from the documentation.
- Additionally: We know of no other open source software package that is built to compute 20-year long term corrected AEP metric from operational data. Practitioners in the field typically write their own code to compute this metric using foundational libraries such as numpy or scikit-learn. There are some companies that offer AEP as part of a commercial software offering, but these products are not open source. I have added a couple sentences around lines 55-58 to clarify this point.

Thank You. Please let me know if you have any other suggestions.
