### 2021-01-22 82ef791
small changes for abbreviations
