+++ paper/paper.md
@@ -81,7 +81,7 @@ OpenOA is an open source framework for operational data analysis of wind energy 
 
 Originally released to the public in 2018 [@osti_1478526], OpenOA is now actively developed through a public Github repository. With over 50 stars on Github, a dozen contributors, and an active issues forum, OpenOA is becoming a mature project that provides a high level interface to solve practical problems in the wind energy industry. OpenOA V2 is released as a Python package and is freely available under a business-friendly, open-source, BSD license. By committing to open source development, OpenOA hopes to facilitate reproducibility of research in this field, provide benchmark implementations of commonly performed data transformation and analysis tasks, and to serve as a conduit that delivers state-of-the-art analysis methods from researchers to practitioners.
 
-Most users will interface with OpenOA through its analysis `methods` module. These are Python classes which conform to a common interface (e.g., they implement `__init__`, `prepare`, and `run` methods). Version 2 of OpenOA implements three high level analysis methods: (1) Long term corrected AEP, (2) electrical losses, and (3) turbine level losses. Uncertainty quantification is achieved in each analysis using a monte carlo approach. A more detailed description of these analyses are provided [in the documentation](https://openoa.readthedocs.io). Low level functions that operate on Pandas series objects are organized in the `toolkit` module. These Python functions are written to be as generic as possible, and can be applied across multiple domains.
+Most users will interface with OpenOA through its analysis `methods` module. These are Python classes which conform to a common interface (e.g., they implement `__init__`, `prepare`, and `run` methods). Version 2 of OpenOA implements three high level analysis methods: (1) Long term corrected annual energy production (AEP), (2) electrical losses, and (3) turbine level losses. Uncertainty quantification is achieved in each analysis using a monte carlo approach. A more detailed description of these analyses are provided [in the documentation](https://openoa.readthedocs.io). Low level functions that operate on Pandas series objects are organized in the `toolkit` module. These Python functions are written to be as generic as possible, and can be applied across multiple domains.
 
 The OpenOA data model is implemented in the `types` module using a class called `PlantData`, which contain at least one Pandas data frame [@Mckinney2010]. These classes add convenience functions and a domain-specific schema based on the IEC 6400-25 standard [@iec25]. OpenOA is part of the ENTR alliance consortium, which envisions a complete software stack centered around an open source implementation of this standard.
 
@@ -91,12 +91,12 @@ OpenOA depends on scikit-learn [@Pedregosa2011] and numpy [@oliphant2006guide], 
 
 # Statement of Need
 
-OpenOA was created by and primarily developed by researchers at the National Renewable Energy Laboratory (NREL) through the Performance, Risk Uncertainty Finance (PRUF) project. The PRUF team recognized the need to compute a 20-year long term correced Annual Energy Production (AEP) metric from operational data as part of an industry-wide benchmarking study [@lunacek2018]. Due to access restrictions on the input data, open source publication of the code was necessary to foster trust in the results of the benchmarking study. Furthermore, after talking with our industry partners, it became clear that there was no industry standard code for computing a 20-year long term corrected AEP. Currently, participants in the wind energy industry who wish to compute metrics like AEP must rely on commercial secondary scada software providers, or must develop their own methodologies internally. We know of no other open source software package to compute 20-year long term corrected AEP.
[…]
