### 2021-01-22 NREL/OpenOA#125 @jordanperr
Remove hard-coded constants
In response to JOSS review, we should place a code review on the roadmap to remove any hard-coded constants that may have crept in over the years.

> I would also caution against too much hardcoding within the code. In some instances, you could make better use of dependency functionality (e.g. operational_analysis/toolkits/unit_conversion.py:convert_power_to_energy could just resample the dataframe rather than expecting one of five hardcoded input time frequencies). In other cases, 'translations' could be kept inside e.g. YAML files, so that they're all in one place, configurable and out of the way of the Python code (e.g. scada_map in examples/project_ENGIE.py).
