### 2021-01-22 #2171 @jordanperr
@brynpickering,

Thank you for the in depth review. We have performed many of the requests you have made. Please review the latest version of the PDF preview from whedon. In particular:

Code:
-	Requirements files were split into examples/requirements.txt and requirements.txt and updated to include all packages required to run the examples.
-	Maximum python version is now 3.8 while we wait for Shapely to update their wheels for Mac.
-	Issue (https://github.com/NREL/OpenOA/issues/121) created to upgrade to a better dependency management system for the project.
-	We have added an issue to refactor the hard-coded constants and instead use those defined in upstream packages. (https://github.com/NREL/OpenOA/issues/125)
-      Gitignore looks OK to me. We specifically allow PNG files due to the embedded graphs in the ReadTheDocs documentation. I would be happy to add more lines to the gitignore if you could be more specific about which files we should add.

Paper:
-	I have integrated all of the smaller style and grammar changes that were suggested.
-	Re-arranged the summary section to incorporate more language about uses of the software earlier on. Combined the abstract with the summary section as suggested.
-	We know of no other open source software package that is built to compute 20-year long term corrected AEP metric from operational data. Practitioners in the field typically write their own code to compute this metric using foundational libraries such as numpy or scikit-learn. There are some companies that offer AEP as part of a commercial software offering, but these products are not open source. I have added a couple sentences around lines 55-58 to clarify this point.
-	Ensure all abbreviations are defined
-	Added a figure with example graphs from the documentation.

Again, thank you so much for your review and your suggestions. Please let me know if there's any further action required.
