### 2021-01-27 3e72306
Updated Nicolas citation, as it now published.
