+++ paper/paper.bib
@@ -45,14 +45,14 @@ year = {2018}
 @article{Bodini2020,
   author={Bodini, Nicola and Optis, Mike},
   title={Are Uncertainty Categories in a Wind Farm Annual Energy Production Estimate Actually Uncorrelated?},
-  journal={Wind Energ. Sci. Discuss.},
-  volume={in review},
-  doi={10.5194/wes-2019-82},
-  url={https://doi.org/10.5194/wes-2019-82},
+  journal={Wind Energ. Sci.},
+  volume={5},
+  pages={1435--1488},
+  doi={10.5194/wes-5-1435-2020},
+  url={https://doi.org/10.5194/wes-5-1435-2020},
   year={2020},
   abstract={Calculations of annual energy production (AEP) from a wind farm – whether based on preconstruction or operational data – are critical for wind farm financial transactions. The uncertainty in the AEP calculation is especially important in quantifying risk and is a key factor in determining financing terms. Standard industry practice assumes that different uncertainty categories within an AEP calculation are uncorrelated and can therefore be combined through a sum of squares approach. In this analysis, we assess the rigor of this assumption by performing operational AEP estimates for over 470 wind farms in the United States. We contrast the standard uncertainty assumption with a Monte Carlo approach to uncertainty quantification in which no assumptions of correlation between uncertainty categories are made. Results show that several uncertainty categories do, in fact, show weak to moderate correlations, namely: wind resource interannual variability and the windiness correction (positive correlation), wind resource interannual variability and regression (negative), and wind speed measurement uncertainty and regression (positive). The sources of these correlations are described and illustrated in detail in this paper, and the effect on the total AEP uncertainty calculation is investigated. Based on these results, we conclude that a Monte Carlo approach to AEP uncertainty quantification is more robust and accurate than the industry standard approach.}
 }
-
 
 
 @book{Rossum2009, 
