### 2021-01-28 ff12c69
Capping supported python version at 3.8 until we can resolve the Shapely dependency issue. (#124)
