+++ test/regression/int_pruf_plant_analysis.py
@@ -227,7 +227,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
  
     def check_simulation_results_lin_monthly(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.41, 2.29, 1.30, 3.57, 0.09, 3.57]
+        expected_results = [12.41, 11, 1.30, 3.57, 0.09, 3.57]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
@@ -240,7 +240,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
 
     def check_simulation_results_lin_daily(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.31, 2.76, 1.36, 4.90, 0.09, 4.91]
+        expected_results = [12.31, 11, 1.36, 4.90, 0.09, 4.91]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
@@ -253,7 +253,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
 
     def check_simulation_results_gam_daily(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.68, 2.50, 1.36, 4.44, 0.087, 4.44]
+        expected_results = [12.68, 11, 1.36, 4.44, 0.087, 4.44]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
@@ -266,7 +266,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
         
     def check_simulation_results_gbm_daily(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.82, 2.84, 1.35, 5.17, 0.09, 5.17]
+        expected_results = [12.82, 11, 1.35, 5.17, 0.09, 5.17]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
@@ -279,7 +279,7 @@ class TestPandasPrufPlantAnalysis(unittest.TestCase):
 
     def check_simulation_results_etr_daily(self, s):
         # Make sure AEP results are consistent to one decimal place
-        expected_results = [12.56, 3.99, 1.35, 5.10, 0.09, 5.10]
+        expected_results = [12.56, 12, 1.35, 5.10, 0.09, 5.10]
 
         calculated_results = [s.aep_GWh.mean(),
                               s.aep_GWh.std() / s.aep_GWh.mean() * 100,
