### 2021-02-02 cb2ed4c
update the dependencies and configs, and show the pre-commit style
