### 2021-02-02 420a6e7
update python requirements for PEP440
