+++ setup.py
@@ -83,5 +83,5 @@ setup(
     install_requires=REQUIRED,
     extras_require=EXTRAS,
     tests_require=TESTS,
-    python_requires="3.6,3.7,3.8",
+    python_requires=">=3.6, !=3.9.*",
 )
