### 2021-02-02 9e414f3
apply pre-commit to selection of code
