+++ operational_analysis/toolkits/unit_conversion.py
@@ -3,7 +3,7 @@ This module provides basic methods for unit conversion and calculation of basic 
 """
 
 
-def convert_power_to_energy(power_col, sample_rate_min='10T'):
+def convert_power_to_energy(power_col, sample_rate_min="10T"):
     """
     Compute energy [kWh] from power [kw] and return the data column
 
@@ -16,16 +16,14 @@ def convert_power_to_energy(power_col, sample_rate_min='10T'):
         :obj:`pandas.Series`: Energy in kWh that matches the length of the input data frame 'df'
 
     """
-    time_conversion = {'1T': 1.,
-                       '5T': 5.,
-                       '10T': 10.,
-                       '30T': 30.,
-                       '1H': 60.}
-    energy_kwh = power_col * time_conversion[sample_rate_min]/ 60.0
+    time_conversion = {"1T": 1.0, "5T": 5.0, "10T": 10.0, "30T": 30.0, "1H": 60.0}
+    energy_kwh = power_col * time_conversion[sample_rate_min] / 60.0
     return energy_kwh
 
 
-def compute_gross_energy(net_energy, avail_losses, curt_losses, avail_type='frac', curt_type='frac'):
+def compute_gross_energy(
+    net_energy, avail_losses, curt_losses, avail_type="frac", curt_type="frac"
+):
     """
     This function computes gross energy for a wind plant or turbine by adding reported availability and
     curtailment losses to reported net energy. Account is made of whether availabilty or curtailment loss data
@@ -41,20 +39,21 @@ def compute_gross_energy(net_energy, avail_losses, curt_losses, avail_type='frac
         gross (numpy array of Pandas series): calculated gross energy for wind plant or turbine
     """
 
-    if (avail_type == 'frac') & (curt_type == 'frac'):
+    if (avail_type == "frac") & (curt_type == "frac"):
         gross = net_energy / (1 - avail_losses - curt_losses)
-    elif (avail_type == 'frac') & (curt_type == 'energy'):
+    elif (avail_type == "frac") & (curt_type == "energy"):
         gross = net_energy / (1 - avail_losses) + curt_losses
-    elif (avail_type == 'energy') & (curt_type == 'frac'):
+    elif (avail_type == "energy") & (curt_type == "frac"):
         gross = net_energy / (1 - curt_losses) + avail_losses
-    elif (avail_type == 'energy') & (curt_type == 'energy'):
+    elif (avail_type == "energy") & (curt_type == "energy"):
         gross = net_energy + curt_losses + avail_losses
 
-    if (len(gross[gross < net_energy]) > 0):
-        raise Exception('Gross energy cannot be less than net energy. Check your input values')
-    if (len(avail_losses[avail_losses < 0]) > 0) | (
-            len(curt_losses[curt_losses < 0]) > 0):
-        raise Exception('Cannot have negative availability or curtailment input values. Check your data')
+    if len(gross[gross < net_energy]) > 0:
+        raise Exception("Gross energy cannot be less than net energy. Check your input values")
+    if (len(avail_losses[avail_losses < 0]) > 0) | (len(curt_losses[curt_losses < 0]) > 0):
+        raise Exception(
+            "Cannot have negative availability or curtailment input values. Check your data"
+        )
 
     return gross
 
