+++ operational_analysis/toolkits/power_curve/functions.py
@@ -1,14 +1,17 @@
-from .parametric_forms import logistic5param
-from .parametric_optimize import fit_parametric_power_curve, least_squares
-from scipy.optimize import differential_evolution
-from pygam import LinearGAM
-import pandas as pd
 import numpy as np
+import pandas as pd
+from pygam import LinearGAM
+from scipy.optimize import differential_evolution
+
+from .parametric_forms import logistic5param
+from .parametric_optimize import least_squares, fit_parametric_power_curve
+
 
 """
 This module holds ready-to-use power curve functions. They take windspeed and power columns as arguments and return a
 python function which can be used to evaluate the power curve at arbitrary locations.
 """
+
 
 def IEC(windspeed_column, power_column, bin_width=0.5, windspeed_start=0, windspeed_end=30.0):
     """
@@ -34,11 +37,11 @@ def IEC(windspeed_column, power_column, bin_width=0.5, windspeed_start=0, windsp
 
     # Compute the mean of each bin and set corresponding P_bin
     for ibin in range(0, len(bins) - 1):
-        indices = ((windspeed_column >= bins[ibin]) & (windspeed_column < bins[ibin + 1]))
+        indices = (windspeed_column >= bins[ibin]) & (windspeed_column < bins[ibin + 1])
         P_bin[ibin] = power_column.loc[indices].mean()
 
     # Linearly interpolate any missing bins
-    P_bin = pd.Series(data=P_bin).interpolate(method='linear').bfill().values
+    P_bin = pd.Series(data=P_bin).interpolate(method="linear").bfill().values
 
     # Create a closure over the computed bins which computes the power curve value for arbitrary array-like input
     def pc_iec(x):
@@ -87,11 +90,14 @@ def logistic_5_parametric(windspeed_column, power_column):
         :obj:`function`: Python function of type (Array[float] -> Array[float]) implementing the power curve.
 
     """
-    return fit_parametric_power_curve(windspeed_column, power_column,
-                                      curve=logistic5param,
-                                      optimization_algorithm=differential_evolution,
-                                      cost_function=least_squares,
-                                      bounds=((1200, 1800), (-10, -1e-3), (1e-3, 30), (1e-3, 1), (1e-3, 10)))
+    return fit_parametric_power_curve(
+        windspeed_column,
+        power_column,
+        curve=logistic5param,
+        optimization_algorithm=differential_evolution,
+        cost_function=least_squares,
+        bounds=((1200, 1800), (-10, -1e-3), (1e-3, 30), (1e-3, 1), (1e-3, 10)),
+    )
 
 
 def gam(windspeed_column, power_column, n_splines=20):
@@ -108,10 +114,7 @@ def gam(windspeed_column, power_column, n_splines=20):
 
     """
     # Fit the model
-    return LinearGAM(n_splines=n_splines).\
-        fit(windspeed_column.values, power_column.values).\
-        predict
-
+    return LinearGAM(n_splines=n_splines).fit(windspeed_column.values, power_column.values).predict
 
 
 def gam_3param(windspeed_column, winddir_column, airdens_column, power_column, n_splines=20):
@@ -130,9 +133,7 @@ def gam_3param(windspeed_column, winddir_column, airdens_column, power_column, n
 
     """
     # create dataframe input to LinearGAM
-    X = pd.DataFrame({'ws': windspeed_column,
-                      'wd': winddir_column,
-                      'dens': airdens_column})
+    X = pd.DataFrame({"ws": windspeed_column, "wd": winddir_column, "dens": airdens_column})
 
     # Set response
     y = power_column.values
@@ -142,8 +143,7 @@ def gam_3param(windspeed_column, winddir_column, airdens_column, power_column, n
 
[…]
