+++ operational_analysis/toolkits/power_curve/parametric_forms.py
@@ -17,11 +17,6 @@ ref: https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.minimiz
 """
 
 
-def _power_curve(x, a, b, c, d, g):
-    """Calculates the power curve value at each point in x."""
-    return [d + (a - d) / (1 + (xx / c) ** b) ** g for xx in x]
-
-
 def logistic5param(x, a, b, c, d, g):
     """Create and return a 5 parameter logistic function
 
@@ -48,7 +43,8 @@ def logistic5param(x, a, b, c, d, g):
         dom = slice(None)  # All elements in domain
 
     # Apply power curve definition to point within domain
-    res[dom] = _power_curve(x[dom], a, b, c, d, g)
+    l5p = lambda xx: d + (a - d) / (1 + (xx / c) ** b) ** g
+    res[dom] = l5p(x[dom])
 
     return res
 
