### 2021-02-02 a3a5b81
removing a lambda per flake8 warning
