### 2021-02-05 #2171 @sjpfenninger
@jordanperr We're ready to accept this. Can you make a tagged release on GitHub and archive that release, for example on Zenodo or figshare, then report the version number and archive DOI back here in the review issue?
