+++ .github/workflows/tags-to-pypi.yml
@@ -10,7 +10,7 @@ jobs:
         name: Build and publish Python package to PyPi
         runs-on: ubuntu-18.04
         steps:
-            - uses: actions/checkout@master
+            - uses: actions/checkout@main
             - name: Set up Python 3.7
               uses: actions/setup-python@v1
               with:
