+++ .github/workflows/tags-to-pypi.yml
@@ -21,7 +21,7 @@ jobs:
                 python setup.py bdist_wheel
                 python setup.py sdist --formats=gztar
             - name: Publish artifacts to PyPI
-              uses: pypa/gh-action-pypi-publish@master
+              uses: pypa/gh-action-pypi-publish@main
               with:
                 password: ${{ secrets.PYPI_PASSWORD }}
                 repository_url: https://upload.pypi.org/legacy/
