### 2021-02-17 5db1113
Feature/master to main (#130)
* updating workflows to reference main branch

* updating language in docs

* missed binder badge in readme
