### 2021-02-18 6b0d4da
updated author list to reflect 2020 and did small grammar fixes for eric and nicola
