+++ paper/paper.md
@@ -8,40 +8,44 @@ tags:
   - standardization
 authors:
   - name: Jordan Perr-Sauer
-    orcid: 0000-0003-0872-7098
+    orcid: 0000-0003-1571-1887
     affiliation: 1 # (Multiple affiliations must be quoted)
   - name: Mike Optis
     orcid: 0000-0001-5617-6134
     affiliation: 1
   - name: Jason M. Fields
     affiliation: 1
+    orcid: 0000-0002-8781-6138
   - name: Nicola Bodini
     orcid: 0000-0002-2550-9853
-    affiliation: 1
-  - name: Lindy Williams
     affiliation: 1
   - name: Joseph C.Y. Lee
     orcid: 0000-0003-1897-6290
     affiliation: 1
   - name: Austin Todd
+    orcid: 0000-0002-1123-0982
     affiliation: 1
-  - name: Caleb Phillips
-    affiliation: 1
-  - name: Monte Lunacek
-    affiliation: 1
-  - name: Anna Craig
-    affiliation: 1
-  - name: Travis Kemper
+  - name: Eric Simley
+    orcid: 0000-0002-1027-9848
     affiliation: 1
   - name: Robert Hammond
     orcid: 0000-0003-4476-6406
+  - name: Caleb Phillips
+    affiliation: 1
+  - name: Monte Lunacek
+    orcid: 0000-0003-3755-224X
+    affiliation: 1
+  - name: Travis Kemper
+    affiliation: 1
+  - name: Lindy Williams
+    affiliation: 1
+  - name: Anna Craig
     affiliation: 1
   - name: Nathan Agarwal
     orcid: 0000-0002-2734-5514
     affiliation: 1
   - name: Shawn Sheng
-    affiliation: 1
-  - name: Eric Simley
+    orcid: 0000-0003-0134-0907
     affiliation: 1
   - name: John Meissner
     affiliation: 1
@@ -77,13 +81,13 @@ References: Is the list of references complete, and is everything cited appropri
 
 # Summary
 
-OpenOA is an open source framework for operational data analysis of wind energy plants, implemented in the Python programming language [@Rossum2009]. OpenOA provides a common data model, high level analysis workflows, and low-level convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate anlytics workflows on operational data sets. OpenOA contains documentation, worked out examples in Jupyter notebooks, and a corresponding example dataset from [Engie Renewable's La Haute Borne Dataset](https://opendata-renewables.engie.com/explore/dataset/d543716b-368d-4c53-8fb1-55addbe8d3ad/information).
+OpenOA is an open source framework for operational data analysis of wind energy plants, implemented in the Python programming language [@Rossum2009]. OpenOA provides a common data model, high level analysis workflows, and low-level convenience functions that engineers, analysts, and researchers in the wind energy industry can use to facilitate anlytics workflows on operational data sets. OpenOA contains documentation, worked out examples in Jupyter notebooks, and a corresponding example dataset from the [Engie Renewable's La Haute Borne Dataset](https://opendata-renewables.engie.com/explore/dataset/d543716b-368d-4c53-8fb1-55addbe8d3ad/information).
 
 Originally released to the public in 2018 [@osti_1478526], OpenOA is now actively developed through a public Github repository. With over 50 stars on Github, a dozen contributors, and an active issues forum, OpenOA is becoming a mature project that provides a high level interface to solve practical problems in the wind energy industry. OpenOA V2 is released as a Python package and is freely available under a business-friendly, open-source, BSD license. By committing to open source development, OpenOA hopes to facilitate reproducibility of research in this field, provide benchmark implementations of commonly performed data transformation and analysis tasks, and to serve as a conduit that delivers state-of-the-art analysis methods from researchers to practitioners.
 
[…]
