+++ paper/paper.bib
@@ -112,8 +112,7 @@ abstractNote = {A collection of software to support the WP3 benchmarking project
 url = {https://www.osti.gov//servlets/purl/1478526},
 doi = {10.11578/dc.20181023.1},
 year = {2018},
-month = {2},
-note = 
+month = {2}
 }
 
 
