+++ paper/paper.bib
@@ -81,7 +81,8 @@ year = {2018}
   volume={445}, 
   pages={51--56}, 
   year={2010}, 
-  organization={Austin, TX} 
+  organization={Austin, TX},
+  doi={10.25080/Majora-92bf1922-00a}
 }
 
 @book{oliphant2006guide, 
@@ -101,7 +102,8 @@ year = {2018}
   number={3}, 
   pages={90--95}, 
   year={2007}, 
-  publisher={IEEE} 
+  publisher={IEEE},
+  doi={10.1109/mcse.2007.55}
 }
 
 
