### 2021-02-18 7de3ba1
Changes to the JOSS paper from peer review (#94)
Updates to JOSS paper from peer review process
