+++ paper/paper.md
@@ -8,31 +8,45 @@ tags:
   - standardization
 authors:
   - name: Jordan Perr-Sauer
-    orcid: 0000-0003-0872-7098
+    orcid: 0000-0003-1571-1887
     affiliation: 1 # (Multiple affiliations must be quoted)
   - name: Mike Optis
     orcid: 0000-0001-5617-6134
     affiliation: 1
   - name: Jason M. Fields
     affiliation: 1
-  - name: Lindy Williams
+    orcid: 0000-0002-8781-6138
+  - name: Nicola Bodini
+    orcid: 0000-0002-2550-9853
     affiliation: 1
   - name: Joseph C.Y. Lee
     orcid: 0000-0003-1897-6290
     affiliation: 1
   - name: Austin Todd
+    orcid: 0000-0002-1123-0982
+    affiliation: 1
+  - name: Eric Simley
+    orcid: 0000-0002-1027-9848
+    affiliation: 1
+  - name: Robert Hammond
+    orcid: 0000-0003-4476-6406
     affiliation: 1
   - name: Caleb Phillips
     affiliation: 1
   - name: Monte Lunacek
-    affiliation: 1
-  - name: Anna Craig
+    orcid: 0000-0003-3755-224X
     affiliation: 1
   - name: Travis Kemper
     affiliation: 1
-  - name: Shuangwen Sheng
+  - name: Lindy Williams
     affiliation: 1
-  - name: Eric Simley
+  - name: Anna Craig
+    affiliation: 1
+  - name: Nathan Agarwal
+    orcid: 0000-0002-2734-5514
+    affiliation: 1
+  - name: Shawn Sheng
+    orcid: 0000-0003-0134-0907
     affiliation: 1
   - name: John Meissner
     affiliation: 1
@@ -43,111 +57,54 @@ date: 20 December 2019
 bibliography: paper.bib
 ---
 
+<!--
+JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that explain the software functionality and domain of use to a non-specialist reader. We also require that authors explain the research applications of the software. The paper should be between 250-1000 words.
+
+Your paper should include:
+
+A list of the authors of the software and their affiliations, using the correct format (see the example below).
+A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
+A clear Statement of Need that illustrates the research purpose of the software.
+A list of key references, including to other software addressing related needs.
+Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
+Acknowledgement of any financial support.
+As this short list shows, JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of Need, Summary, Acknowledgements, and References sections. You can look at an example accepted paper. Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.
+
+
+Review Checklist:
+Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
+A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
+State of the field: Do the authors describe how this software compares to other commonly-used packages?
+Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
+References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
+
+-->
+
 # Summary
 
-OpenOA is an open source Python package which implements operational analysis (OA) methods for wind energy plants.
[…]
