+++ paper/paper.bib
@@ -109,12 +109,13 @@ year = {2018}
 
 @misc{osti_1478526,
 title = {Open OA, FKA: Wind Plant Performance Project (WP3) Benchmarking},
-author = {Perr-Sauer, Jordan and Fields, Michael and Craig, Anna and Optis, Michael and Kemper, Travis and Sheng, Shuangwen and Meissner, John and Phillips, Caleb and USDOE Office of Energy Efficiency and Renewable Energy},
+author = {Perr-Sauer, Jordan and Fields, Michael and Craig, Anna and Optis, Michael and Kemper, Travis and Sheng, Shuangwen and Meissner, John and Phillips, Caleb},
 abstractNote = {A collection of software to support the WP3 benchmarking project, including (but not limited to) operational analysis, historical validation, and benchmarking code.},
 url = {https://www.osti.gov//servlets/purl/1478526},
 doi = {10.11578/dc.20181023.1},
 year = {2018},
-month = {2}
+month = {2},
+publisher = {USDOE Office of Energy Efficiency and Renewable Energy}
 }
 
 
