### 2021-02-19 #2171 @jordanperr
@sjpfenninger I think we are finally done! Everything has been merged into the main branch and the version for publication in JOSS has been tagged "v2.1"

Repo: https://github.com/NREL/OpenOA
Paper Branch: main
Tag: v2.1
Zenodo DOI: 10.5281/zenodo.4549876

We made a few last minute tweaks to the paper, which are listed here:

- Changed the order of some authors to reflect the new reality of the contributions (as of v2.1). The old author order would have been more relevant 1 year ago, when this submission was started.
- Removed reference to "20-year long term corrected AEP" throughout the paper in favor of "long term corrected AEP", which is more correct.
- Instead of claiming there is no industry standard code for long term corrected AEP, we updated the language to reflect that there is no industry standard method, either. This is a stronger claim than just code.
- Minor changes to the grammar.
- Fixed citations so that Whedon clears them all.
