### 2021-02-23 #2171 @kthyng
@jordanperr I can help move this publication forward from here since it looks about finished! Can you update the metadata on your zenodo archive so that the author list and title exactly match your JOSS paper? Also you mentioned a CI issue in your comment — would you like to wait to finish this until that is fixed?
