### 2021-02-23 #2171 @kthyng
@jordanperr I read through your paper and it looks good pending some corrections you can check out in this [PR](https://github.com/NREL/OpenOA/pull/131).
