### 2021-02-26 #2171 @kthyng
Congratulations on your new publication to @jordanperr and team!! Many thanks to editor @sjpfenninger and reviewers @gschivley and @brynpickering — we couldn't make JOSS happen without your time and expertise!! 🎉
